% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.r
\name{a_aes}
\alias{a_aes}
\title{Define aesthetic mappings.}
\usage{
a_aes(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics can be omitted (because
they are so common); all other aesthetics must be named.}
}
\description{
Generate aesthetic mappings that describe how variables in the data are
mapped to visual properties (aesthetics) of geoms. This function also
standardise aesthetic names by performs partial name matching, converting
color to colour, and old style R names to a_plot names (eg. pch to shape,
cex to size)
}
\examples{
a_aes(x = mpg, y = wt)
a_aes(mpg, wt)

# You can also map aesthetics to functions of variables
a_aes(x = mpg ^ 2, y = wt / cyl)

# Aesthetic names are automatically standardised
a_aes(col = x)
a_aes(fg = x)
a_aes(color = x)
a_aes(colour = x)

# a_aes is almost always used with a_plot() or a layer
a_plot(mpg, a_aes(displ, hwy)) + a_geom_point()
a_plot(mpg) + a_geom_point(a_aes(displ, hwy))

# a_aesthetics supplied to a_plot() are used as defaults for every layer
# you can override them, or supply different aesthetics for each layer
}
\seealso{
See \code{\link{a_aes_q}}/\code{\link{a_aes_string}} for standard
  evaluation versions of \code{a_aes}.

See
   \code{\link{a_aes_colour_fill_alpha}}, \code{\link{a_aes_group_order}},
   \code{\link{a_aes_linetype_size_shape}} and \code{\link{a_aes_position}}
   for more specific examples with different aesthetics.
}
