% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.r
\name{a_aes_}
\alias{a_aes_}
\alias{a_aes_string}
\alias{a_aes_q}
\title{Define aesthetic mappings from strings, or quoted calls and formulas.}
\usage{
a_aes_(x, y, ...)

a_aes_string(x, y, ...)

a_aes_q(x, y, ...)
}
\arguments{
\item{x, y, ...}{List of name value pairs. Elements must be either
quoted calls, strings, one-sided formulas or constants.}
}
\description{
Aesthetic mappings describe how variables in the data are mapped to visual
properties (aesthetics) of geoms. \code{\link{a_aes}} uses non-standard
evaluation to capture the variable names. \code{a_aes_} and \code{a_aes_string}
require you to explicitly quote the inputs either with \code{""} for
\code{a_aes_string()}, or with \code{quote} or \code{~} for \code{a_aes_()}.
(\code{a_aes_q} is an alias to \code{a_aes_})
}
\details{
It's better to use \code{a_aes_q()}, because there's no easy way to create the
equivalent to \code{a_aes(colour = "my colour")} or \code{a_aes{x = `X$1`}}
with \code{a_aes_string()}.

\code{a_aes_string} and \code{a_aes_} are particularly useful when writing
functions that create plots because you can use strings or quoted
names/calls to define the aesthetic mappings, rather than having to use
\code{\link{substitute}} to generate a call to \code{a_aes()}.
}
\examples{
# Three ways of generating the same aesthetics
a_aes(mpg, wt, col = cyl)
a_aes_(quote(mpg), quote(wt), col = quote(cyl))
a_aes_(~mpg, ~wt, col = ~cyl)
a_aes_string("mpg", "wt", col = "cyl")

# You can't easily mimic these calls with a_aes_string
a_aes(`$100`, colour = "smooth")
a_aes_(~ `$100`, colour = "smooth")
# Ok, you can, but it requires a _lot_ of quotes
a_aes_string("`$100`", colour = '"smooth"')

# Convert strings to names with as.name
var <- "cyl"
a_aes(col = x)
a_aes_(col = as.name(var))
}
\seealso{
\code{\link{a_aes}}
}
