% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-map.r
\name{a_annotation_map}
\alias{a_annotation_map}
\title{Annotation: maps.}
\usage{
a_annotation_map(map, ...)
}
\arguments{
\item{map}{data frame representing a map.  Most map objects can be
converted into the right format by using \code{\link{a_fortify}}}

\item{...}{other arguments used to modify aesthetics}
}
\description{
Annotation: maps.
}
\examples{
if (require("maps")) {
usamap <- map_data("state")

seal.sub <- subset(seals, long > -130 & lat < 45 & lat > 40)
a_plot(seal.sub, a_aes(x = long, y = lat)) +
  a_annotation_map(usamap, fill = "NA", colour = "grey50") +
  a_geom_segment(a_aes(xend = long + delta_long, yend = lat + delta_lat))

seal2 <- transform(seal.sub,
  latr = cut(lat, 2),
  longr = cut(long, 2))

a_plot(seal2,  a_aes(x = long, y = lat)) +
  a_annotation_map(usamap, fill = "NA", colour = "grey50") +
  a_geom_segment(a_aes(xend = long + delta_long, yend = lat + delta_lat)) +
  a_facet_grid(latr ~ longr, scales = "free", space = "free")
}
}
