% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bench.r
\name{a_benchplot}
\alias{a_benchplot}
\title{Benchmark plot creation time.
Broken down into construct, build, render and draw times.}
\usage{
a_benchplot(x)
}
\arguments{
\item{x}{code to create ggplot2 plot}
}
\description{
Benchmark plot creation time.
Broken down into construct, build, render and draw times.
}
\examples{
a_benchplot(a_plot(mtcars, a_aes(mpg, wt)) + a_geom_point())
a_benchplot(a_plot(mtcars, a_aes(mpg, wt)) + a_geom_point() + a_facet_grid(. ~ cyl))
}
\keyword{internal}
