% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{a_build_strip}
\alias{a_build_strip}
\title{a_build strip function}
\usage{
a_build_strip(panel, label_df, labeller, a_theme, side = "right",
  switch = NULL)
}
\arguments{
\item{panel}{....}

\item{label_df}{....}

\item{labeller}{....}

\item{a_theme}{.....}

\item{side}{.....}

\item{switch}{....}
}
\description{
a_build strip function
}
