% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-flip.r
\name{a_coord_flip}
\alias{a_coord_flip}
\title{Flipped cartesian coordinates.}
\usage{
a_coord_flip(xlim = NULL, ylim = NULL, expand = TRUE)
}
\arguments{
\item{xlim}{Limits for the x and y axes.}

\item{ylim}{Limits for the x and y axes.}

\item{expand}{If \code{TRUE}, the default, adds a small expansion factor to
the limits to ensure that data and axes don't overlap. If \code{FALSE},
limits are taken exactly from the data or \code{xlim}/\code{ylim}.}
}
\description{
Flipped cartesian coordinates so that horizontal becomes vertical, and
vertical, horizontal. This is primarily useful for converting geoms and
statistics which display y conditional on x, to x conditional on y.
}
\examples{
# Very useful for creating boxplots, and other interval
# geoms in the horizontal instead of vertical position.

a_plot(diamonds, a_aes(cut, price)) +
  a_geom_boxplot() +
  a_coord_flip()

h <- a_plot(diamonds, a_aes(carat)) +
  a_geom_histogram()
h
h + a_coord_flip()
h + a_coord_flip() + a_scale_x_reverse()

# You can also use it to flip line and area plots:
df <- data.frame(x = 1:5, y = (1:5) ^ 2)
a_plot(df, a_aes(x, y)) +
  a_geom_area()
last_plot() + a_coord_flip()
}
