% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{a_element_grob}
\alias{a_element_grob}
\title{Generate grid grob from theme a_element}
\usage{
a_element_grob(a_element, ...)
}
\arguments{
\item{a_element}{Theme a_element, i.e. \code{a_element_rect} or similar.}

\item{...}{Other arguments to control specific of rendering. This is
usually at least position. See the source code for individual methods.}
}
\description{
Generate grid grob from theme a_element
}
\keyword{internal}
