% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{a_element_line}
\alias{a_element_line}
\title{Theme a_element: line.}
\usage{
a_element_line(colour = NULL, size = NULL, linetype = NULL,
  lineend = NULL, color = NULL)
}
\arguments{
\item{colour}{line colour}

\item{size}{line size}

\item{linetype}{line type}

\item{lineend}{line end}

\item{color}{an alias for \code{colour}}
}
\description{
Theme a_element: line.
}
