% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-elements.r
\name{a_element_rect}
\alias{a_element_rect}
\title{Theme a_element: rectangle.}
\usage{
a_element_rect(fill = NULL, colour = NULL, size = NULL, linetype = NULL,
  color = NULL)
}
\arguments{
\item{fill}{fill colour}

\item{colour}{border colour}

\item{size}{border size}

\item{linetype}{border linetype}

\item{color}{an alias for \code{colour}}
}
\description{
Most often used for backgrounds and borders.
}
