% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.r
\name{a_fortify}
\alias{a_fortify}
\title{Fortify a model with data.}
\usage{
a_fortify(model, data, ...)
}
\arguments{
\item{model}{model or other R object to convert to data frame}

\item{data}{original dataset, if needed}

\item{...}{other arguments passed to methods}
}
\description{
Rather than using this function, I now recomend using the \pkg{broom}
package, which implements a much wider range of methods. \code{a_fortify}
may be deprecated in the future.
}
\seealso{
\code{\link{a_fortify.lm}}
}
