% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-contour.r, R/stat-contour.r
\name{a_geom_contour}
\alias{a_geom_contour}
\alias{a_stat_contour}
\title{Display contours of a 3d surface in 2d.}
\usage{
a_geom_contour(mapping = NULL, data = NULL, a_stat = "contour",
  a_position = "identity", ..., lineend = "butt", linejoin = "round",
  linemitre = 1, na.rm = FALSE, show.legend = NA, inherit.a_aes = TRUE)

a_stat_contour(mapping = NULL, data = NULL, a_geom = "contour",
  a_position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{a_aes}} or
\code{\link{a_aes_}}. If specified and \code{inherit.a_aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{a_plot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{a_fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{a_stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{a_position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{a_layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired a_geom/a_stat.}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}

\item{linemitre}{Line mitre limit (number greater than 1)}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.a_aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{a_geom}{The geometric object to use display the data}
}
\description{
Display contours of a 3d surface in 2d.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{animint2:::rd_aesthetics("a_geom", "contour")}
}

\section{Computed variables}{

\describe{
 \item{level}{height of contour}
}
}

\examples{
#' # Basic plot
v <- a_plot(faithfuld, a_aes(waiting, eruptions, z = density))
v + a_geom_contour()

# Or compute from raw data
a_plot(faithful, a_aes(waiting, eruptions)) +
  a_geom_density_2d()

\donttest{
# Setting bins creates evenly spaced contours in the range of the data
v + a_geom_contour(bins = 2)
v + a_geom_contour(bins = 10)

# Setting binwidth does the same thing, parameterised by the distance
# between contours
v + a_geom_contour(binwidth = 0.01)
v + a_geom_contour(binwidth = 0.001)

# Other parameters
v + a_geom_contour(a_aes(colour = ..level..))
v + a_geom_contour(colour = "red")
v + a_geom_raster(a_aes(fill = density)) +
  a_geom_contour(colour = "white")
}
}
\seealso{
\code{\link{a_geom_density_2d}}: 2d density contours
}
