% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-count.r, R/stat-sum.r
\name{a_geom_count}
\alias{a_geom_count}
\alias{a_stat_sum}
\title{Count the number of observations at each location.}
\usage{
a_geom_count(mapping = NULL, data = NULL, a_stat = "sum",
  a_position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.a_aes = TRUE)

a_stat_sum(mapping = NULL, data = NULL, a_geom = "point",
  a_position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{a_aes}} or
\code{\link{a_aes_}}. If specified and \code{inherit.a_aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{a_plot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{a_fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{a_position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{a_layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired a_geom/a_stat.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.a_aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{a_geom, a_stat}{Use to override the default connection between
\code{a_geom_count} and \code{a_stat_sum}.}
}
\description{
This is a variant \code{\link{a_geom_point}} that counts the number of
observations at each location, then maps the count to point size. It
useful when you have discrete data.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{animint2:::rd_aesthetics("a_geom", "point")}
}

\section{Computed variables}{

\describe{
 \item{n}{number of observations at a_position}
 \item{prop}{percent of points in that panel at that a_position}
}
}

\examples{
a_plot(mpg, a_aes(cty, hwy)) +
 a_geom_point()

a_plot(mpg, a_aes(cty, hwy)) +
 a_geom_count()

# Best used in conjunction with a_scale_size_area which ensures that
# counts of zero would be given size 0. Doesn't make much different
# here because the smallest count is already close to 0.
a_plot(mpg, a_aes(cty, hwy)) +
 a_geom_count()
 a_scale_size_area()

# Display proportions instead of counts -------------------------------------
# By default, all categorical variables in the plot form the groups.
# Specifying a_geom_count without a group identifier leads to a plot which is
# not useful:
d <- a_plot(diamonds, a_aes(x = cut, y = clarity))
d + a_geom_count(a_aes(size = ..prop..))
# To correct this problem and achieve a more desirable plot, we need
# to specify which group the proportion is to be calculated over.
d + a_geom_count(a_aes(size = ..prop.., group = 1)) +
  a_scale_size_area(max_size = 10)

# Or group by x/y variables to have rows/columns sum to 1.
d + a_geom_count(a_aes(size = ..prop.., group = cut)) +
  a_scale_size_area(max_size = 10)
d + a_geom_count(a_aes(size = ..prop.., group = clarity)) +
  a_scale_size_area(max_size = 10)
}
