% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-density2d.r, R/stat-density-2d.r
\name{a_geom_density_2d}
\alias{a_geom_density_2d}
\alias{a_geom_density2d}
\alias{a_stat_density_2d}
\alias{a_stat_density2d}
\title{Contours from a 2d density estimate.}
\usage{
a_geom_density_2d(mapping = NULL, data = NULL, a_stat = "density2d",
  a_position = "identity", ..., lineend = "butt", linejoin = "round",
  linemitre = 1, na.rm = FALSE, show.legend = NA, inherit.a_aes = TRUE)

a_stat_density_2d(mapping = NULL, data = NULL, a_geom = "density_2d",
  a_position = "identity", ..., contour = TRUE, n = 100, h = NULL,
  na.rm = FALSE, show.legend = NA, inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{a_aes}} or
\code{\link{a_aes_}}. If specified and \code{inherit.a_aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{a_plot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{a_fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{a_position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{a_layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired a_geom/a_stat.}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}

\item{linemitre}{Line mitre limit (number greater than 1)}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.a_aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{a_geom, a_stat}{Use to override the default connection between
\code{a_geom_density_2d} and \code{a_stat_density_2d}.}

\item{contour}{If \code{TRUE}, contour the results of the 2d density
estimation}

\item{n}{number of grid points in each direction}

\item{h}{Bandwidth (vector of length two). If \code{NULL}, estimated
using \code{\link[MASS]{bandwidth.nrd}}.}
}
\description{
Perform a 2D kernel density estimation using kde2d and display the
results with contours. This can be useful for dealing with overplotting.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{animint2:::rd_aesthetics("a_geom", "density_2d")}
}

\section{Computed variables}{

Same as \code{\link{a_stat_contour}}
}

\examples{
m <- a_plot(faithful, a_aes(x = eruptions, y = waiting)) +
 a_geom_point() +
 xlim(0.5, 6) +
 ylim(40, 110)
m + a_geom_density_2d()
\donttest{
m + a_stat_density_2d(a_aes(fill = ..level..), a_geom = "polygon")

set.seed(4393)
dsmall <- diamonds[sample(nrow(diamonds), 1000), ]
d <- a_plot(dsmall, a_aes(x, y))
# If you map an a_aesthetic to a categorical variable, you will get a
# set of contours for each value of that variable
d + a_geom_density_2d(a_aes(colour = cut))

# If we turn contouring off, we can use use geoms like tiles:
d + a_stat_density_2d(a_geom = "raster", a_aes(fill = ..density..), contour = FALSE)
# Or points:
d + a_stat_density_2d(a_geom = "point", a_aes(size = ..density..), n = 20, contour = FALSE)
}
}
\seealso{
\code{\link{a_geom_contour}} for contour drawing a_geom,
 \code{\link{a_stat_sum}} for another way of dealing with overplotting
}
