% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-jitter.r
\name{a_geom_jitter}
\alias{a_geom_jitter}
\title{Points, jittered to reduce overplotting.}
\usage{
a_geom_jitter(mapping = NULL, data = NULL, a_stat = "identity",
  a_position = "jitter", ..., width = NULL, height = NULL,
  na.rm = FALSE, show.legend = NA, inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{a_aes}} or
\code{\link{a_aes_}}. If specified and \code{inherit.a_aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{a_plot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{a_fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{a_stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{a_position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{a_layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired a_geom/a_stat.}

\item{width}{Amount of vertical and horizontal jitter. The jitter
  is added in both positive and negative directions, so the total spread
  is twice the value specified here.

  If omitted, defaults to 40\% of the resolution of the data: this means the
  jitter values will occupy 80\% of the implied bins. Categorical data
  is aligned on the integers, so a width or height of 0.5 will spread the
  data so it's not possible to see the distinction between the categories.}

\item{height}{Amount of vertical and horizontal jitter. The jitter
  is added in both positive and negative directions, so the total spread
  is twice the value specified here.

  If omitted, defaults to 40\% of the resolution of the data: this means the
  jitter values will occupy 80\% of the implied bins. Categorical data
  is aligned on the integers, so a width or height of 0.5 will spread the
  data so it's not possible to see the distinction between the categories.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.a_aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
}
\description{
The jitter geom is a convenient default for a_geom_point with a_position =
'jitter'. It's a useful way of handling overplotting caused by discreteness
in smaller datasets.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{animint2:::rd_aesthetics("a_geom", "point")}
}

\examples{
p <- a_plot(mpg, a_aes(cyl, hwy))
p + a_geom_point()
p + a_geom_jitter()

# Add aesthetic mappings
p + a_geom_jitter(a_aes(colour = class))

# Use smaller width/height to emphasise categories
a_plot(mpg, a_aes(cyl, hwy)) + a_geom_jitter()
a_plot(mpg, a_aes(cyl, hwy)) + a_geom_jitter(width = 0.25)

# Use larger width/height to completely smooth away discreteness
a_plot(mpg, a_aes(cty, hwy)) + a_geom_jitter()
a_plot(mpg, a_aes(cty, hwy)) + a_geom_jitter(width = 0.5, height = 0.5)
}
\seealso{
\code{\link{a_geom_point}} for regular, unjittered points,
 \code{\link{a_geom_boxplot}} for another way of looking at the conditional
    distribution of a variable
}
