% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-quantile.r, R/stat-quantile.r
\name{a_geom_quantile}
\alias{a_geom_quantile}
\alias{a_stat_quantile}
\title{Add quantile lines from a quantile regression.}
\usage{
a_geom_quantile(mapping = NULL, data = NULL, a_stat = "quantile",
  a_position = "identity", ..., lineend = "butt", linejoin = "round",
  linemitre = 1, na.rm = FALSE, show.legend = NA, inherit.a_aes = TRUE)

a_stat_quantile(mapping = NULL, data = NULL, a_geom = "quantile",
  a_position = "identity", ..., quantiles = c(0.25, 0.5, 0.75),
  formula = NULL, method = "rq", method.args = list(), na.rm = FALSE,
  show.legend = NA, inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{a_aes}} or
\code{\link{a_aes_}}. If specified and \code{inherit.a_aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{a_plot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{a_fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{a_position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{a_layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired a_geom/a_stat.}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}

\item{linemitre}{Line mitre limit (number greater than 1)}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.a_aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{a_geom, a_stat}{Use to override the default connection between
\code{a_geom_quantile} and \code{a_stat_quantile}.}

\item{quantiles}{conditional quantiles of y to calculate and display}

\item{formula}{formula relating y variables to x variables}

\item{method}{Quantile regression method to use.  Currently only supports
\code{\link[quantreg]{rq}}.}

\item{method.args}{List of additional arguments passed on to the modelling
function defined by \code{method}.}
}
\description{
This can be used as a continuous analogue of a a_geom_boxplot.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{animint2:::rd_aesthetics("a_geom", "quantile")}
}

\section{Computed variables}{

\describe{
  \item{quantile}{quantile of distribution}
}
}

\examples{
m <- a_plot(mpg, a_aes(displ, 1 / hwy)) + a_geom_point()
m + a_geom_quantile()
m + a_geom_quantile(quantiles = 0.5)
q10 <- seq(0.05, 0.95, by = 0.05)
m + a_geom_quantile(quantiles = q10)

# You can also use rqss to fit smooth quantiles
m + a_geom_quantile(method = "rqss")
# Note that rqss doesn't pick a smoothing constant automatically, so
# you'll need to tweak lambda yourself
m + a_geom_quantile(method = "rqss", lambda = 0.1)

# Set aesthetics to fixed value
m + a_geom_quantile(colour = "red", size = 2, alpha = 0.5)
}
