% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_geoms.R
\name{a_geom_tallrect}
\alias{a_geom_tallrect}
\title{ggplot2 a_geom with xmin and xmax aesthetics that covers the entire y range, useful for clickSelects background elements.}
\usage{
a_geom_tallrect(mapping = NULL, data = NULL, a_stat = "identity",
  a_position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data set}

\item{a_stat}{statistic mapping, defaults to identity}

\item{a_position}{position mapping, defaults to identity}

\item{...}{other arguments}

\item{na.rm}{...}

\item{show.legend}{...}

\item{inherit.a_aes}{...}
}
\value{
ggplot2 layer
}
\description{
ggplot2 a_geom with xmin and xmax aesthetics that covers the entire y range, useful for clickSelects background elements.
}
\examples{
library(animint2)

## Example: 2 plots, 2 selectors, but only interacting with 1 plot.
data(breakpoints)
only.error <- subset(breakpoints$error,type=="E")
only.segments <- subset(only.error,bases.per.probe==bases.per.probe[1])
signal.colors <- c(estimate="#0adb0a",
                   latent="#0098ef")
breakpointError <- 
  list(signal=a_plot()+
         a_geom_point(a_aes(position, signal),
                    showSelected="bases.per.probe",
                    data=breakpoints$signals)+
         a_geom_line(a_aes(position, signal), colour=signal.colors[["latent"]],
                   data=breakpoints$imprecision)+
         a_geom_segment(a_aes(first.base, mean, xend=last.base, yend=mean),
                      showSelected=c("segments", "bases.per.probe"),
                      colour=signal.colors[["estimate"]],
                      data=breakpoints$segments)+
         a_geom_vline(a_aes(xintercept=base),
                    showSelected=c("segments", "bases.per.probe"),
                    colour=signal.colors[["estimate"]],
                    linetype="dashed",
                    data=breakpoints$breaks),
       error=a_plot()+
         a_geom_vline(a_aes(xintercept=segments), clickSelects="segments",
                    data=only.segments, lwd=17, alpha=1/2)+
         a_geom_line(a_aes(segments, error, group=bases.per.probe),
                   clickSelects="bases.per.probe",
                   data=only.error, lwd=4))
animint2dir(breakpointError)
}
