% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_geoms.R
\name{a_geom_widerect}
\alias{a_geom_widerect}
\title{ggplot2 geom with ymin and ymax aesthetics that covers the entire x range, useful for clickSelects background elements.}
\usage{
a_geom_widerect(mapping = NULL, data = NULL, a_stat = "identity",
  a_position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data set}

\item{a_stat}{statistic mapping, defaults to identity}

\item{a_position}{position mapping, defaults to identity}

\item{...}{other arguments}

\item{na.rm}{...}

\item{show.legend}{...}

\item{inherit.a_aes}{...}
}
\value{
ggplot2 layer
}
\description{
ggplot2 geom with ymin and ymax aesthetics that covers the entire x range, useful for clickSelects background elements.
}
\examples{
 \dontrun{ 
   source(system.file("examples/WorldBank.R", package = "animint"))
 }
}
