% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-axis.r
\name{a_guide_axis}
\alias{a_guide_axis}
\title{Grob for axes}
\usage{
a_guide_axis(at, a_labels, a_position = "right", a_theme)
}
\arguments{
\item{at}{...}

\item{a_labels}{at ticks}

\item{a_position}{of ticks}

\item{a_theme}{...}
}
\description{
Grob for axes
}
