% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-colorbar.r
\name{a_guide_colourbar}
\alias{a_guide_colourbar}
\alias{a_guide_colorbar}
\title{Continuous colour bar guide.}
\usage{
a_guide_colourbar(title = waiver(), title.a_position = NULL,
  title.a_theme = NULL, title.hjust = NULL, title.vjust = NULL,
  a_label = TRUE, a_label.a_position = NULL, a_label.a_theme = NULL,
  a_label.hjust = NULL, a_label.vjust = NULL, barwidth = NULL,
  barheight = NULL, nbin = 20, raster = TRUE, ticks = TRUE,
  draw.ulim = TRUE, draw.llim = TRUE, direction = NULL,
  default.unit = "line", reverse = FALSE, order = 0, ...)

a_guide_colorbar(title = waiver(), title.a_position = NULL,
  title.a_theme = NULL, title.hjust = NULL, title.vjust = NULL,
  a_label = TRUE, a_label.a_position = NULL, a_label.a_theme = NULL,
  a_label.hjust = NULL, a_label.vjust = NULL, barwidth = NULL,
  barheight = NULL, nbin = 20, raster = TRUE, ticks = TRUE,
  draw.ulim = TRUE, draw.llim = TRUE, direction = NULL,
  default.unit = "line", reverse = FALSE, order = 0, ...)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link{waiver}}), the name of the scale object or the name
specified in \code{\link{labs}} is used for the title.}

\item{title.a_position}{A character string indicating the a_position of a
 title. One of "top" (default for a vertical a_guide), "bottom", "left"
(default for a horizontal a_guide), or "right."}

\item{title.a_theme}{A a_theme object for rendering the title text. Usually the
object of \code{\link{a_element_text}} is expected. By default, the a_theme is
specified by \code{legend.title} in \code{\link{a_theme}} or a_theme.}

\item{title.hjust}{A number specifying horizontal justification of the
title text.}

\item{title.vjust}{A number specifying vertical justification of the title
text.}

\item{a_label}{logical. If \code{TRUE} then the labels are drawn. If
\code{FALSE} then the labels are invisible.}

\item{a_label.a_position}{A character string indicating the a_position of a
a_label. One of "top", "bottom" (default for horizontal guide), "left", or
"right" (default for vertical guide).}

\item{a_label.a_theme}{A a_theme object for rendering the a_label text. Usually the
object of \code{\link{a_element_text}} is expected. By default, the a_theme is
specified by \code{legend.text} in \code{\link{a_theme}} or a_theme.}

\item{a_label.hjust}{A numeric specifying horizontal justification of the
a_label text.}

\item{a_label.vjust}{A numeric specifying vertical justification of the a_label
text.}

\item{barwidth}{A numeric or a \code{\link[grid]{unit}} object specifying
the width of the colorbar. Default value is \code{legend.key.width} or
\code{legend.key.size} in \code{\link{a_theme}} or a_theme.}

\item{barheight}{A numeric or a \code{\link[grid]{unit}} object specifying
the height of the colorbar. Default value is \code{legend.key.height} or
\code{legend.key.size} in \code{\link{a_theme}} or a_theme.}

\item{nbin}{A numeric specifying the number of bins for drawing colorbar. A
smoother colorbar for a larger value.}

\item{raster}{A logical. If \code{TRUE} then the colorbar is rendered as a
raster object. If \code{FALSE} then the colorbar is rendered as a set of
rectangles. Note that not all graphics devices are capable of rendering
raster image.}

\item{ticks}{A logical specifying if tick marks on colorbar should be
visible.}

\item{draw.ulim}{A logical specifying if the upper limit tick marks should
be visible.}

\item{draw.llim}{A logical specifying if the lower limit tick marks should
be visible.}

\item{direction}{A character string indicating the direction of the guide.
One of "horizontal" or "vertical."}

\item{default.unit}{A character string indicating \code{\link[grid]{unit}}
for \code{barwidth} and \code{barheight}.}

\item{reverse}{logical. If \code{TRUE} the colorbar is reversed. By default,
the highest value is on the top and the lowest value is on the bottom}

\item{order}{positive integer less that 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{...}{ignored.}
}
\value{
A guide object
}
\description{
Colour bar guide shows continuous color scales mapped onto values.
Colour bar is available with \code{a_scale_fill} and \code{a_scale_colour}.
For more information, see the inspiration for this function:
\href{http://www.mathworks.com/help/techdoc/ref/colorbar.html}{Matlab's colorbar function}.
}
\details{
Guides can be specified in each \code{a_scale_*} or in \code{\link{a_guides}}.
\code{a_guide="legend"} in \code{a_scale_*} is syntactic sugar for
\code{a_guide=a_guide_legend()} (e.g. \code{a_scale_color_manual(a_guide = "legend")}).
As for how to specify the a_guide for each scale in more detail,
see \code{\link{a_guides}}.
}
\examples{
df <- reshape2::melt(outer(1:4, 1:4), varnames = c("X1", "X2"))

p1 <- a_plot(df, a_aes(X1, X2)) + a_geom_tile(a_aes(fill = value))
p2 <- p1 + a_geom_point(a_aes(size = value))

# Basic form
p1 + a_scale_fill_continuous(a_guide = "colorbar")
p1 + a_scale_fill_continuous(a_guide = a_guide_colorbar())
p1 + a_guides(fill = a_guide_colorbar())

# Control styles

# bar size
p1 + a_guides(fill = a_guide_colorbar(barwidth = 0.5, barheight = 10))

# no a_label
p1 + a_guides(fill = a_guide_colorbar(a_label = FALSE))

# no tick marks
p1 + a_guides(fill = a_guide_colorbar(ticks = FALSE))

# a_label position
p1 + a_guides(fill = a_guide_colorbar(a_label.a_position = "left"))

# a_label a_theme
p1 + a_guides(fill = a_guide_colorbar(a_label.a_theme = a_element_text(colour = "blue", angle = 0)))

# small number of bins
p1 + a_guides(fill = a_guide_colorbar(nbin = 3))

# large number of bins
p1 + a_guides(fill = a_guide_colorbar(nbin = 100))

# make top- and bottom-most ticks invisible
p1 + a_scale_fill_continuous(limits = c(0,20), breaks = c(0, 5, 10, 15, 20),
 a_guide = a_guide_colorbar(nbin=100, draw.ulim = FALSE, draw.llim = FALSE))

# guides can be controlled independently
p2 +
  a_scale_fill_continuous(a_guide = "colorbar") +
  a_scale_size(a_guide = "legend")
p2 + a_guides(fill = "colorbar", size = "legend")

p2 +
  a_scale_fill_continuous(a_guide = a_guide_colorbar(direction = "horizontal")) +
  a_scale_size(a_guide = a_guide_legend(direction = "vertical"))
}
\seealso{
Other a_guides: \code{\link{a_guide_legend}},
  \code{\link{a_guides}}
}
