% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide-legend.r
\name{a_guide_legend}
\alias{a_guide_legend}
\title{Legend guide.}
\usage{
a_guide_legend(title = waiver(), title.a_position = NULL,
  title.a_theme = NULL, title.hjust = NULL, title.vjust = NULL,
  a_label = TRUE, a_label.a_position = NULL, a_label.a_theme = NULL,
  a_label.hjust = NULL, a_label.vjust = NULL, keywidth = NULL,
  keyheight = NULL, direction = NULL, default.unit = "line",
  override.a_aes = list(), nrow = NULL, ncol = NULL, byrow = FALSE,
  reverse = FALSE, order = 0, ...)
}
\arguments{
\item{title}{A character string or expression indicating a title of guide.
If \code{NULL}, the title is not shown. By default
(\code{\link{waiver}}), the name of the scale object or the name
specified in \code{\link{labs}} is used for the title.}

\item{title.a_position}{A character string indicating the a_position of a
 title. One of "top" (default for a vertical a_guide), "bottom", "left"
(default for a horizontal a_guide), or "right."}

\item{title.a_theme}{A a_theme object for rendering the title text. Usually the
object of \code{\link{a_element_text}} is expected. By default, the a_theme is
specified by \code{legend.title} in \code{\link{a_theme}} or a_theme.}

\item{title.hjust}{A number specifying horizontal justification of the
title text.}

\item{title.vjust}{A number specifying vertical justification of the title
text.}

\item{a_label}{logical. If \code{TRUE} then the labels are drawn. If
\code{FALSE} then the labels are invisible.}

\item{a_label.a_position}{A character string indicating the a_position of a
a_label. One of "top", "bottom" (default for horizontal guide), "left", or
"right" (default for vertical guide).}

\item{a_label.a_theme}{A a_theme object for rendering the a_label text. Usually the
object of \code{\link{a_element_text}} is expected. By default, the a_theme is
specified by \code{legend.text} in \code{\link{a_theme}} or a_theme.}

\item{a_label.hjust}{A numeric specifying horizontal justification of the
a_label text.}

\item{a_label.vjust}{A numeric specifying vertical justification of the a_label
text.}

\item{keywidth}{A numeric or a \code{\link[grid]{unit}} object specifying
the width of the legend key. Default value is \code{legend.key.width} or
\code{legend.key.size} in \code{\link{a_theme}} or a_theme.}

\item{keyheight}{A numeric or a \code{\link[grid]{unit}} object specifying
the height of the legend key. Default value is \code{legend.key.height} or
\code{legend.key.size} in \code{\link{a_theme}} or a_theme.}

\item{direction}{A character string indicating the direction of the guide.
One of "horizontal" or "vertical."}

\item{default.unit}{A character string indicating \code{\link[grid]{unit}}
for \code{keywidth} and \code{keyheight}.}

\item{override.a_aes}{A list specifying aesthetic parameters of legend key.
See details and examples.}

\item{nrow}{The desired number of rows of legends.}

\item{ncol}{The desired number of column of legends.}

\item{byrow}{logical. If \code{FALSE} (the default) the legend-matrix is
filled by columns, otherwise the legend-matrix is filled by rows.}

\item{reverse}{logical. If \code{TRUE} the order of legends is reversed.}

\item{order}{positive integer less that 99 that specifies the order of
this guide among multiple guides. This controls the order in which
multiple guides are displayed, not the contents of the guide itself.
If 0 (default), the order is determined by a secret algorithm.}

\item{...}{ignored.}
}
\value{
A guide object
}
\description{
Legend type guide shows key (i.e., geoms) mapped onto values.
Legend guides for various scales are integrated if possible.
}
\details{
Guides can be specified in each \code{a_scale_*} or in \code{\link{a_guides}}.
\code{a_guide="legend"} in \code{a_scale_*} is syntactic sugar for
\code{a_guide=a_guide_legend()} (e.g. \code{a_scale_color_manual(a_guide = "legend")}).
As for how to specify the a_guide for each scale in more detail,
see \code{\link{a_guides}}.
}
\examples{
\donttest{
df <- reshape2::melt(outer(1:4, 1:4), varnames = c("X1", "X2"))

p1 <- a_plot(df, a_aes(X1, X2)) + a_geom_tile(a_aes(fill = value))
p2 <- p1 + a_geom_point(a_aes(size = value))

# Basic form
p1 + a_scale_fill_continuous(a_guide = "legend")
p1 + a_scale_fill_continuous(a_guide = a_guide_legend())

# Guide title
p1 + a_scale_fill_continuous(a_guide = a_guide_legend(title = "V")) # title text
p1 + a_scale_fill_continuous(a_guide = a_guide_legend(title = NULL)) # no title

# Control styles

# key size
p1 + a_guides(fill = a_guide_legend(keywidth = 3, keyheight = 1))

# title position
p1 + a_guides(fill = a_guide_legend(title = "LEFT", title.a_position = "left"))

# title text styles via a_element_text
p1 + a_guides(fill =
  a_guide_legend(
    title.a_theme = a_element_text(
      size = 15,
      face = "italic",
      colour = "red",
      angle = 0
    )
  )
)

# a_label position
p1 + a_guides(fill = a_guide_legend(a_label.a_position = "left", a_label.hjust = 1))

# a_label styles
p1 + a_scale_fill_continuous(breaks = c(5, 10, 15),
  a_labels = paste("long", c(5, 10, 15)),
  a_guide = a_guide_legend(
    direction = "horizontal",
    title.a_position = "top",
    a_label.a_position = "bottom",
    a_label.hjust = 0.5,
    a_label.vjust = 1,
    a_label.a_theme = a_element_text(angle = 90)
  )
)

# Set aesthetic of legend key

# very low alpha value make it difficult to see legend key
p3 <- a_plot(diamonds, a_aes(carat, price)) +
  a_geom_point(a_aes(colour = color), alpha = 1/100)
p3

# override.a_aes overwrites the alpha
p3 + a_guides(colour = a_guide_legend(override.a_aes = list(alpha = 1)))

# multiple row/col legends
df <- data.frame(x = 1:20, y = 1:20, color = letters[1:20])
p <- a_plot(df, a_aes(x, y)) +
  a_geom_point(a_aes(colour = color))
p + a_guides(col = a_guide_legend(nrow = 8))
p + a_guides(col = a_guide_legend(ncol = 8))
p + a_guides(col = a_guide_legend(nrow = 8, byrow = TRUE))
p + a_guides(col = a_guide_legend(ncol = 8, byrow = TRUE))

# reversed order legend
p + a_guides(col = a_guide_legend(reverse = TRUE))
}
}
\seealso{
Other a_guides: \code{\link{a_guide_colourbar}},
  \code{\link{a_guides}}
}
