% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{a_guides}
\alias{a_guides}
\title{Set guides for each scale.}
\usage{
a_guides(...)
}
\arguments{
\item{...}{List of scale guide pairs}
}
\value{
A list containing the mapping between scale and guide.
}
\description{
Guides for each scale can be set in call of \code{a_scale_*} with argument
\code{a_guide}, or in \code{a_guides}.
}
\examples{
\donttest{
# ggplot object

dat <- data.frame(x = 1:5, y = 1:5, p = 1:5, q = factor(1:5),
 r = factor(1:5))
p <- a_plot(dat, a_aes(x, y, colour = p, size = q, shape = r)) + a_geom_point()

# without guide specification
p

# Show colorbar guide for colour.
# All these examples below have a same effect.

p + a_guides(colour = "colorbar", size = "legend", shape = "legend")
p + a_guides(colour = a_guide_colorbar(), size = a_guide_legend(),
  shape = a_guide_legend())
p +
 a_scale_colour_continuous(a_guide = "colorbar") +
 a_scale_size_discrete(a_guide = "legend") +
 a_scale_shape(a_guide = "legend")

 # Remove some a_guides
 p + a_guides(colour = "none")
 p + a_guides(colour = "colorbar",size = "none")

# Guides are integrated where possible

p + a_guides(colour = a_guide_legend("title"), size = a_guide_legend("title"),
  shape = a_guide_legend("title"))
# same as
g <- a_guide_legend("title")
p + a_guides(colour = g, size = g, shape = g)

p + a_theme(legend.a_position = "bottom")

# a_position of a_guides

p + a_theme(legend.a_position = "bottom", legend.box = "horizontal")

# Set order for multiple guides
a_plot(mpg, a_aes(displ, cty)) +
  a_geom_point(a_aes(size = hwy, colour = cyl, shape = drv)) +
  a_guides(
   colour = a_guide_colourbar(order = 1),
   shape = a_guide_legend(order = 2),
   size = a_guide_legend(order = 3)
 )
}
}
\seealso{
Other a_guides: \code{\link{a_guide_colourbar}},
  \code{\link{a_guide_legend}}
}
