% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{a_guides_build}
\alias{a_guides_build}
\title{build up all guide boxes into one guide-boxes.}
\usage{
a_guides_build(ggrobs, a_theme)
}
\arguments{
\item{ggrobs}{...}

\item{a_theme}{...}
}
\description{
build up all guide boxes into one guide-boxes.
}
