% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{a_guides_geom}
\alias{a_guides_geom}
\title{a_guides_geom function}
\usage{
a_guides_geom(gdefs, layers, default_mapping)
}
\arguments{
\item{gdefs}{...}

\item{layers}{....}

\item{default_mapping}{...}
}
\description{
a_guides_geom function
}
