% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{a_guides_train}
\alias{a_guides_train}
\title{train each scale in scales and generate the definition of guide}
\usage{
a_guides_train(scales, a_theme, a_guides, a_labels)
}
\arguments{
\item{scales}{...}

\item{a_theme}{...}

\item{a_guides}{...}

\item{a_labels}{....}
}
\description{
train each scale in scales and generate the definition of guide
}
