% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-labels.r
\name{a_label_bquote}
\alias{a_label_bquote}
\title{Backquoted labeller}
\usage{
a_label_bquote(rows = NULL, cols = NULL, default = a_label_value)
}
\arguments{
\item{rows}{Backquoted labelling expression for rows.}

\item{cols}{Backquoted labelling expression for columns.}

\item{default}{Default labeller function for the rows or the
columns when no plotmath expression is provided.}
}
\description{
\code{\link{a_label_bquote}()} offers a flexible way of labelling
facet rows or columns with plotmath expressions. Backquoted
variables will be replaced with their value in the facet.
}
\examples{
# The variables mentioned in the plotmath expression must be
# backquoted and referred to by their names.
p <- a_plot(mtcars, a_aes(wt, mpg)) + a_geom_point()
p + a_facet_grid(vs ~ ., labeller = a_label_bquote(alpha ^ .(vs)))
p + a_facet_grid(. ~ vs, labeller = a_label_bquote(cols = .(vs) ^ .(vs)))
p + a_facet_grid(. ~ vs + am, labeller = a_label_bquote(cols = .(am) ^ .(vs)))
}
\seealso{
\link{labellers}, \code{\link{labeller}()},
}
