% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-summary.r
\name{a_mean_se}
\alias{a_mean_se}
\title{Calculate mean and standard errors on either side.}
\usage{
a_mean_se(x, mult = 1)
}
\arguments{
\item{x}{numeric vector}

\item{mult}{number of multiples of standard error}
}
\description{
Calculate mean and standard errors on either side.
}
\seealso{
for use with \code{\link{a_stat_summary}}
}
