% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.r
\name{a_plot_build}
\alias{a_plot_build}
\alias{a_layer_data}
\alias{a_layer_scales}
\alias{a_layer_grob}
\title{Build a_plot for rendering.}
\usage{
a_plot_build(plot)

a_layer_data(plot, i = 1L)

a_layer_scales(plot, i = 1L, j = 1L)

a_layer_grob(plot, i = 1L)
}
\arguments{
\item{plot}{a_plot object}
}
\description{
\code{a_plot_build} takes the plot object, and performs all steps necessary
to produce an object that can be rendered.  This function outputs two pieces:
a list of data frames (one for each layer), and a panel object, which
contain all information about axis limits, breaks etc.
}
\details{
\code{a_layer_data}, \code{a_layer_grob}, and \code{a_layer_scales} are helper
functions that returns the data, grob, or scales associated with a given
layer. These are useful for tests.
}
\seealso{
\code{\link{print.a_plot}} and \code{\link{benchplot}} for
 functions that contain the complete set of steps for generating
 a a_plot2 plot.
}
\keyword{internal}
