% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-identity.r
\name{a_position_identity}
\alias{a_position_identity}
\title{Don't adjust position}
\usage{
a_position_identity()
}
\description{
Don't adjust position
}
\seealso{
Other position adjustments: \code{\link{a_position_dodge}},
  \code{\link{a_position_fill}},
  \code{\link{a_position_jitterdodge}},
  \code{\link{a_position_jitter}},
  \code{\link{a_position_nudge}}
}
