% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-jitterdodge.R
\name{a_position_jitterdodge}
\alias{a_position_jitterdodge}
\title{Adjust position by simultaneously dodging and jittering}
\usage{
a_position_jitterdodge(jitter.width = NULL, jitter.height = 0,
  dodge.width = 0.75)
}
\arguments{
\item{jitter.width}{degree of jitter in x direction. Defaults to 40\% of the
resolution of the data.}

\item{jitter.height}{degree of jitter in y direction. Defaults to 0.}

\item{dodge.width}{the amount to dodge in the x direction. Defaults to 0.75,
the default \code{a_position_dodge()} width.}
}
\description{
This is primarily used for aligning points generated through
\code{a_geom_point()} with dodged boxplots (e.g., a \code{a_geom_boxplot()} with
a fill aesthetic supplied).
}
\examples{
dsub <- diamonds[ sample(nrow(diamonds), 1000), ]
a_plot(dsub, a_aes(x = cut, y = carat, fill = clarity)) +
  a_geom_boxplot(outlier.size = 0) +
  a_geom_point(pch = 21, a_position = a_position_jitterdodge())
}
\seealso{
Other position adjustments: \code{\link{a_position_dodge}},
  \code{\link{a_position_fill}},
  \code{\link{a_position_identity}},
  \code{\link{a_position_jitter}},
  \code{\link{a_position_nudge}}
}
