% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-resolution.r
\name{a_resolution}
\alias{a_resolution}
\title{Compute the "a_resolution" of a data vector.}
\usage{
a_resolution(x, zero = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{zero}{should a zero value be automatically included in the
computation of a_resolution}
}
\description{
The a_resolution is is the smallest non-zero distance between adjacent
values.  If there is only one unique value, then the a_resolution is defined
to be one.
}
\details{
If x is an integer vector, then it is assumed to represent a discrete
variable, and the a_resolution is 1.
}
\examples{
a_resolution(1:10)
a_resolution((1:10) - 0.5)
a_resolution((1:10) - 0.5, FALSE)
a_resolution(c(1,2, 10, 20, 50))
a_resolution(as.integer(c(1, 10, 20, 50)))  # Returns 1
}
