% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-alpha.r
\name{a_scale_alpha}
\alias{a_scale_alpha}
\alias{a_scale_alpha_continuous}
\alias{a_scale_alpha_discrete}
\title{Alpha scales.}
\usage{
a_scale_alpha(..., range = c(0.1, 1))

a_scale_alpha_continuous(..., range = c(0.1, 1))

a_scale_alpha_discrete(..., range = c(0.1, 1))
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{continuous_a_scale}}
or \code{\link{discrete_a_scale}} as appropriate, to control name, limits,
breaks, labels and so forth.}

\item{range}{range of output alpha values.  Should lie between 0 and 1.}
}
\description{
\code{a_scale_alpha} is an alias for \code{a_scale_alpha_continuous} since
that is the most common use of alpha, and it saves a bit of typing.
}
\examples{
(p <- a_plot(mtcars, a_aes(mpg, cyl)) +
  a_geom_point(a_aes(alpha = cyl)))
p + a_scale_alpha("cylinders")
p + a_scale_alpha("number\\nof\\ncylinders")

p + a_scale_alpha(range = c(0.4, 0.8))

(p <- a_plot(mtcars, a_aes(mpg, cyl)) +
  a_geom_point(a_aes(alpha = factor(cyl))))
p + a_scale_alpha_discrete(range = c(0.4, 0.8))
}
