% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-discrete-.r
\name{a_scale_x_discrete}
\alias{a_scale_x_discrete}
\alias{a_scale_y_discrete}
\title{Discrete position.}
\usage{
a_scale_x_discrete(..., expand = waiver())

a_scale_y_discrete(..., expand = waiver())
}
\arguments{
\item{...}{common discrete scale parameters: \code{name}, \code{breaks},
\code{a_labels}, \code{na.value}, \code{limits} and \code{a_guide}.  See
\code{\link{discrete_a_scale}} for more details}

\item{expand}{a numeric vector of length two giving multiplicative and
additive expansion constants. These constants ensure that the data is
placed some distance away from the axes.}
}
\description{
You can use continuous positions even with a discrete position scale -
this allows you (e.g.) to place labels between bars in a bar chart.
Continuous positions are numeric values starting at one for the first
level, and increasing by one for each level (i.e. the labels are placed
at integer positions).  This is what allows jittering to work.
}
\examples{
a_plot(diamonds, a_aes(cut)) + a_geom_bar()

\donttest{
# The discrete position scale is added automatically whenever you
# have a discrete position.

(d <- a_plot(subset(diamonds, carat > 1), a_aes(cut, clarity)) +
      a_geom_jitter())

d + a_scale_x_discrete("Cut")
d + a_scale_x_discrete("Cut", a_labels = c("Fair" = "F","Good" = "G",
  "Very Good" = "VG","Perfect" = "P","Ideal" = "I"))

# Use limits to adjust the which levels (and in what order)
# are displayed
d + a_scale_x_discrete(limits = c("Fair","Ideal"))

# you can also use the short hand functions xlim and ylim
d + xlim("Fair","Ideal", "Good")
d + ylim("I1", "IF")

# See ?reorder to reorder based on the values of another variable
a_plot(mpg, a_aes(manufacturer, cty)) + a_geom_point()
a_plot(mpg, a_aes(reorder(manufacturer, cty), cty)) + a_geom_point()
a_plot(mpg, a_aes(reorder(manufacturer, displ), cty)) + a_geom_point()

# Use abbreviate as a formatter to reduce long names
a_plot(mpg, a_aes(reorder(manufacturer, displ), cty)) +
  a_geom_point() +
  a_scale_x_discrete(a_labels = abbreviate)
}
}
