% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-grey.r, R/uu_zxx.r
\name{a_scale_colour_grey}
\alias{a_scale_colour_grey}
\alias{a_scale_fill_grey}
\alias{a_scale_color_grey}
\title{Sequential grey colour scale.}
\usage{
a_scale_colour_grey(..., start = 0.2, end = 0.8, na.value = "red")

a_scale_fill_grey(..., start = 0.2, end = 0.8, na.value = "red")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_a_scale}}
to control name, limits, breaks, labels and so forth.}

\item{start}{gray value at low end of palette}

\item{end}{gray value at high end of palette}

\item{na.value}{Colour to use for missing values}
}
\description{
Based on \code{\link{gray.colors}}
}
\examples{
p <- a_plot(mtcars, a_aes(mpg, wt)) + a_geom_point(a_aes(colour = factor(cyl)))
p + a_scale_colour_grey()
p + a_scale_colour_grey(end = 0)

# You may want to turn off the pale grey background with this scale
p + a_scale_colour_grey() + a_theme_bw()

# Colour of missing values is controlled with na.value:
miss <- factor(sample(c(NA, 1:5), nrow(mtcars), replace = TRUE))
a_plot(mtcars, a_aes(mpg, wt)) +
  a_geom_point(a_aes(colour = miss)) +
  a_scale_colour_grey()
a_plot(mtcars, a_aes(mpg, wt)) +
  a_geom_point(a_aes(colour = miss)) +
  a_scale_colour_grey(na.value = "green")
}
\seealso{
Other colour scales:
  \code{\link{a_scale_colour_brewer}},
  \code{\link{a_scale_colour_gradient}},
  \code{\link{a_scale_colour_hue}}
}
