% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-identity.r, R/uu_zxx.r
\name{a_scale_identity}
\alias{a_scale_identity}
\alias{a_scale_colour_identity}
\alias{a_scale_fill_identity}
\alias{a_scale_shape_identity}
\alias{a_scale_linetype_identity}
\alias{a_scale_alpha_identity}
\alias{a_scale_size_identity}
\alias{a_scale_color_identity}
\title{Use values without scaling.}
\usage{
a_scale_colour_identity(..., guide = "none")

a_scale_fill_identity(..., guide = "none")

a_scale_shape_identity(..., guide = "none")

a_scale_linetype_identity(..., guide = "none")

a_scale_alpha_identity(..., guide = "none")

a_scale_size_identity(..., guide = "none")
}
\arguments{
\item{...}{Other arguments passed on to \code{\link{discrete_a_scale}} or
\code{\link{continuous_a_scale}}}

\item{guide}{Guide to use for this scale - defaults to \code{"none"}.}
}
\description{
Use values without scaling.
}
\examples{
a_plot(luv_colours, aes(u, v)) +
  a_geom_point(aes(colour = col), size = 3) +
  a_scale_color_identity() +
  a_coord_equal()

df <- data.frame(
  x = 1:4,
  y = 1:4,
  colour = c("red", "green", "blue", "yellow")
)
a_plot(df, aes(x, y)) + a_geom_tile(aes(fill = colour))
a_plot(df, aes(x, y)) +
  a_geom_tile(aes(fill = colour)) +
  a_scale_fill_identity()

# To get a legend guide, specify guide = "legend"
a_plot(df, aes(x, y)) +
  a_geom_tile(aes(fill = colour)) +
  a_scale_fill_identity(guide = "legend")
# But you'll typically also need to supply breaks and labels:
a_plot(df, aes(x, y)) +
  a_geom_tile(aes(fill = colour)) +
  a_scale_fill_identity("trt", labels = letters[1:4], breaks = df$colour,
  guide = "legend")

# cyl scaled to appropriate size
a_plot(mtcars, aes(mpg, wt)) + a_geom_point(aes(size = cyl))

# cyl used as point size
a_plot(mtcars, aes(mpg, wt)) +
  a_geom_point(aes(size = cyl)) +
  a_scale_size_identity()
}
