% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-shape.r
\name{a_scale_shape}
\alias{a_scale_shape}
\alias{a_scale_shape_discrete}
\alias{a_scale_shape_continuous}
\title{Scale for shapes, aka glyphs.}
\usage{
a_scale_shape(..., solid = TRUE)
}
\arguments{
\item{...}{common discrete scale parameters: \code{name}, \code{breaks},
\code{a_labels}, \code{na.value}, \code{limits} and \code{a_guide}.  See
\code{\link{discrete_a_scale}} for more details}

\item{solid}{Are the shapes solid, \code{TRUE}, or hollow \code{FALSE}?}
}
\description{
A continuous variable can not be mapped to shape.
}
\examples{
dsmall <- diamonds[sample(nrow(diamonds), 100), ]

(d <- a_plot(dsmall, a_aes(carat, price)) + a_geom_point(a_aes(shape = cut)))
d + a_scale_shape(solid = TRUE) # the default
d + a_scale_shape(solid = FALSE)
d + a_scale_shape(name = "Cut of diamond")
d + a_scale_shape(name = "Cut of\\ndiamond")

# To change order of levels, change order of
# underlying factor
levels(dsmall$cut) <- c("Fair", "Good", "Very Good", "Premium", "Ideal")

# Need to recreate plot to pick up new data
a_plot(dsmall, a_aes(price, carat)) + a_geom_point(a_aes(shape = cut))

# Or for short:
d \%+\% dsmall
}
