% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales-.r
\name{a_scales_list}
\alias{a_scales_list}
\title{Scales object encapsulates multiple scales.
All input and output done with data.frames to facilitate
multiple input and output variables}
\usage{
a_scales_list()
}
\description{
Scales object encapsulates multiple scales.
All input and output done with data.frames to facilitate
multiple input and output variables
}
