% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-path.r
\name{a_stairstep}
\alias{a_stairstep}
\title{Calculate a_stairsteps
Used by \code{\link{a_geom_step}}}
\usage{
a_stairstep(data, direction = "hv")
}
\arguments{
\item{data}{...}

\item{direction}{...}
}
\description{
Calculate a_stairsteps
Used by \code{\link{a_geom_step}}
}
