% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-function.r
\name{a_stat_function}
\alias{a_stat_function}
\title{Superimpose a function.}
\usage{
a_stat_function(mapping = NULL, data = NULL, a_geom = "path",
  a_position = "identity", ..., fun, xlim = NULL, n = 101,
  args = list(), na.rm = FALSE, show.legend = NA, inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{a_aes}} or
\code{\link{a_aes_}}. If specified and \code{inherit.a_aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{a_plot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{a_fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{a_geom}{The geometric object to use display the data}

\item{a_position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{a_layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired a_geom/a_stat.}

\item{fun}{function to use}

\item{xlim}{Optionally, restrict the range of the function to this range.}

\item{n}{number of points to interpolate along}

\item{args}{list of additional arguments to pass to \code{fun}}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.a_aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
}
\description{
Superimpose a function.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{animint2:::rd_aesthetics("a_stat", "function")}
}

\section{Computed variables}{

\describe{
  \item{x}{x's along a grid}
  \item{y}{value of function evaluated at corresponding x}
}
}

\examples{
set.seed(1492)
df <- data.frame(
  x = rnorm(100)
)
x <- df$x
base <- a_plot(df, a_aes(x)) + a_geom_density()
base + a_stat_function(fun = dnorm, colour = "red")
base + a_stat_function(fun = dnorm, colour = "red", args = list(mean = 3))

# Plot functions without data
# Examples adapted from Kohske Takahashi

# Specify range of x-axis
a_plot(data.frame(x = c(0, 2)), a_aes(x)) +
  a_stat_function(fun = exp, a_geom = "line")

# Plot a normal curve
a_plot(data.frame(x = c(-5, 5)), a_aes(x)) + a_stat_function(fun = dnorm)

# To specify a different mean or sd, use the args parameter to supply new values
a_plot(data.frame(x = c(-5, 5)), a_aes(x)) +
  a_stat_function(fun = dnorm, args = list(mean = 2, sd = .5))

# Two functions on the same plot
f <- a_plot(data.frame(x = c(0, 10)), a_aes(x))
f + a_stat_function(fun = sin, colour = "red") +
  a_stat_function(fun = cos, colour = "blue")

# Using a custom function
test <- function(x) {x ^ 2 + x + 20}
f + a_stat_function(fun = test)
}
