% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-summary-bin.R, R/stat-summary.r
\name{a_stat_summary_bin}
\alias{a_stat_summary_bin}
\alias{a_stat_summary}
\title{Summarise y values at unique/binned x x.}
\usage{
a_stat_summary_bin(mapping = NULL, data = NULL, a_geom = "pointrange",
  a_position = "identity", ..., fun.data = NULL, fun.y = NULL,
  fun.ymax = NULL, fun.ymin = NULL, fun.args = list(), na.rm = FALSE,
  show.legend = NA, inherit.a_aes = TRUE)

a_stat_summary(mapping = NULL, data = NULL, a_geom = "pointrange",
  a_position = "identity", ..., fun.data = NULL, fun.y = NULL,
  fun.ymax = NULL, fun.ymin = NULL, fun.args = list(), na.rm = FALSE,
  show.legend = NA, inherit.a_aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{a_aes}} or
\code{\link{a_aes_}}. If specified and \code{inherit.a_aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{a_plot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{a_fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{a_geom}{Use to override the default connection between
\code{a_geom_histogram}/\code{a_geom_freqpoly} and \code{a_stat_bin}.}

\item{a_position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{other arguments passed on to \code{\link{a_layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired a_geom/a_stat.}

\item{fun.data}{A function that is given the complete data and should
return a data frame with variables \code{ymin}, \code{y}, and \code{ymax}.}

\item{fun.ymin, fun.y, fun.ymax}{Alternatively, supply three individual
functions that are each passed a vector of x's and should return a
single number.}

\item{fun.args}{Optional additional arguments passed on to the functions.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.a_aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}
}
\description{
\code{a_stat_summary} operates on unique \code{x}; \code{a_stat_summary_bin}
operators on binned \code{x}. They are more flexible versions of
\code{\link{a_stat_bin}}: instead of just counting, they can compute any
aggregate.
}
\section{Aesthetics}{

\Sexpr[results=rd,stage=build]{animint2:::rd_aesthetics("a_stat", "summary")}
}

\section{Summary functions}{

You can either supply summary functions individually (\code{fun.y},
\code{fun.ymax}, \code{fun.ymin}), or as a single function (\code{fun.data}):

\describe{
  \item{fun.data}{Complete summary function. Should take numeric vector as
     input and return data frame as output}
  \item{fun.ymin}{ymin summary function (should take numeric vector and
    return single number)}
  \item{fun.y}{y summary function (should take numeric vector and return
    single number)}
  \item{fun.ymax}{ymax summary function (should take numeric vector and
    return single number)}
}

A simple vector function is easiest to work with as you can return a single
number, but is somewhat less flexible. If your summary function computes
multiple values at once (e.g. ymin and ymax), use \code{fun.data}.

If no aggregation functions are suppled, will default to
\code{\link{a_mean_se}}.
}

\examples{
d <- a_plot(mtcars, a_aes(cyl, mpg)) + a_geom_point()

# You can supply individual functions to summarise the value at
# each x:
d + a_stat_summary(fun.y = "median", colour = "red", size = 2, a_geom = "point")
d + a_stat_summary(fun.y = "mean", colour = "red", size = 2, a_geom = "point")
d + a_aes(colour = factor(vs)) + a_stat_summary(fun.y = mean, a_geom="line")

d + a_stat_summary(fun.y = mean, fun.ymin = min, fun.ymax = max,
  colour = "red")

d <- a_plot(diamonds, a_aes(cut))
d + a_geom_bar()
d + a_stat_summary_bin(a_aes(y = price), fun.y = "mean", a_geom = "bar")

\donttest{
# Don't use ylim to zoom into a summary plot - this throws the
# data away
p <- a_plot(mtcars, a_aes(cyl, mpg)) +
  a_stat_summary(fun.y = "mean", a_geom = "point")
p
p + ylim(15, 30)
# Instead use a_coord_cartesian
p + a_coord_cartesian(ylim = c(15, 30))

# A set of useful summary functions is provided from the Hmisc package:
a_stat_sum_df <- function(fun, a_geom="crossbar", ...) {
  a_stat_summary(fun.data = fun, colour = "red", a_geom = a_geom, width = 0.2, ...)
}
d <- a_plot(mtcars, a_aes(cyl, mpg)) + a_geom_point()
# The crossbar a_geom needs grouping to be specified when used with
# a continuous x axis.
d + a_stat_sum_df("a_mean_cl_boot", mapping = a_aes(group = cyl))
d + a_stat_sum_df("a_mean_sdl", mapping = a_aes(group = cyl))
d + a_stat_sum_df("a_mean_sdl", fun.args = list(mult = 1), mapping = a_aes(group = cyl))
d + a_stat_sum_df("a_median_hilow", mapping = a_aes(group = cyl))

# An example with highly skewed distributions:
if (require("ggplot2movies")) {
set.seed(596)
mov <- movies[sample(nrow(movies), 1000), ]
 m2 <- a_plot(mov, a_aes(x = factor(round(rating)), y = votes)) + a_geom_point()
 m2 <- m2 + a_stat_summary(fun.data = "a_mean_cl_boot", a_geom = "crossbar",
                         colour = "red", width = 0.3) + xlab("rating")
m2
# Notice how the overplotting skews off visual perception of the mean
# supplementing the raw data with summary statistics is _very_ important

# Next, we'll look at votes on a log scale.

# Transforming the scale means the data are transformed
# first, after which statistics are computed:
m2 + a_scale_y_log10()
# Transforming the coordinate system occurs after the
# statistic has been computed. This means we're calculating the summary on the raw data
# and stretching the a_geoms onto the log scale.  Compare the widths of the
# standard errors.
m2 + a_coord_trans(y="log10")
}
}
}
\seealso{
\code{\link{a_geom_errorbar}}, \code{\link{a_geom_pointrange}},
 \code{\link{a_geom_linerange}}, \code{\link{a_geom_crossbar}} for geoms to
 display summarised data
}
