% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-defaults.r
\name{aa_theme}
\alias{aa_theme}
\alias{a_theme_grey}
\alias{a_theme_gray}
\alias{a_theme_bw}
\alias{a_theme_linedraw}
\alias{a_theme_light}
\alias{a_theme_minimal}
\alias{a_theme_classic}
\alias{a_theme_dark}
\alias{a_theme_void}
\title{ggplot2 a_themes}
\usage{
a_theme_grey(base_size = 11, base_family = "")

a_theme_gray(base_size = 11, base_family = "")

a_theme_bw(base_size = 12, base_family = "")

a_theme_linedraw(base_size = 12, base_family = "")

a_theme_light(base_size = 12, base_family = "")

a_theme_minimal(base_size = 12, base_family = "")

a_theme_classic(base_size = 12, base_family = "")

a_theme_dark(base_size = 12, base_family = "")

a_theme_void(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
a_themes set the general aspect of the plot such as the colour of the
background, gridlines, the size and colour of fonts.
}
\details{
\describe{

\item{\code{a_theme_gray}}{
The signature ggplot2 a_theme with a grey background and white gridlines,
designed to put the data forward yet make comparisons easy.}

\item{\code{a_theme_bw}}{
The classic dark-on-light ggplot2 a_theme. May work better for presentations
displayed with a projector.}

\item{\code{a_theme_linedraw}}{
A a_theme with only black lines of various widths on white backgrounds,
reminiscent of a line drawings. Serves a purpose similar to \code{a_theme_bw}.
Note that this a_theme has some very thin lines (<< 1 pt) which some journals
may refuse.}

\item{\code{a_theme_light}}{
A a_theme similar to \code{a_theme_linedraw} but with light grey lines and axes,
to direct more attention towards the data.}

\item{\code{a_theme_dark}}{
The dark cousin of \code{a_theme_light}, with similar line sizes but a dark background. Useful to make thin coloured lines pop out.}

\item{\code{a_theme_minimal}}{
A minimalistic a_theme with no background annotations.}

\item{\code{a_theme_classic}}{
A classic-looking a_theme, with x and y axis lines and no gridlines.}

\item{\code{a_theme_void}}{
A completely empty a_theme.}

}
}
\examples{
p <- a_plot(mtcars) + a_geom_point(a_aes(x = wt, y = mpg,
     colour = factor(gear))) + a_facet_wrap(~am)

p
p + a_theme_gray()
p + a_theme_bw()
p + a_theme_linedraw()
p + a_theme_light()
p + a_theme_dark()
p + a_theme_minimal()
p + a_theme_classic()
p + a_theme_void()

}
