% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grob-absolute.r
\name{absoluteGrob}
\alias{absoluteGrob}
\title{Absolute grob
This grob has fixed dimensions and position.
It's still experimental}
\usage{
absoluteGrob(grob, width = NULL, height = NULL, xmin = NULL,
  ymin = NULL, vp = NULL)
}
\arguments{
\item{grob}{....}

\item{width}{.....}

\item{height}{.....}

\item{xmin}{......}

\item{ymin}{.......}

\item{vp}{....}
}
\description{
Absolute grob
This grob has fixed dimensions and position.
It's still experimental
}
