% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{addShowSelectedForLegend}
\alias{addShowSelectedForLegend}
\title{Add a showSelected aesthetic if legend is specified}
\usage{
addShowSelectedForLegend(meta, legend, L)
}
\arguments{
\item{meta}{meta object with all information}

\item{legend}{legend to scan for showSelected}

\item{L}{layer of the plot}
}
\value{
L : Layer with additional mapping to new aesthetic
}
\description{
Add a showSelected aesthetic if legend is specified
}
