% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{addSSandCSasAesthetics}
\alias{addSSandCSasAesthetics}
\title{Add the showSelected/clickSelects params to the aesthetics mapping}
\usage{
addSSandCSasAesthetics(a_aesthetics, extra_params)
}
\arguments{
\item{a_aesthetics}{list. Original aesthetics mapping of the layer}

\item{extra_params}{named list containing the details of showSelected
and clickSelects values of the layer}
}
\value{
Modified aesthetics list with showSelected/clickSelects params added
}
\description{
Add the showSelected/clickSelects params to the aesthetics mapping
}
\details{
Used before calling ggplot_build in parsePlot and while checking
animint extensions to raise error
}
