% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animint.R
\name{animint2dir}
\alias{animint2dir}
\alias{animint}
\title{Compile and render an animint in a local directory}
\usage{
animint2dir(plot.list, out.dir = tempfile(), json.file = "plot.json",
  open.browser = interactive(), css.file = "")
}
\arguments{
\item{plot.list}{a named list of ggplots and option lists.}

\item{out.dir}{directory to store html/js/csv files.}

\item{json.file}{character string that names the JSON file with metadata associated with the plot.}

\item{open.browser}{Should R open a browser? If yes, be sure to configure your browser to allow access to local files, as some browsers block this by default (e.g. chrome).}

\item{css.file}{character string for non-empty css file to include. Provided file will be copied to the output directory as styles.css}
}
\value{
invisible list of ggplots in list format.
}
\description{
An animint is a list of ggplots and options that defines
an interactive animation and can be viewed in a web browser.
Several new aesthetics control interactivity.
The most important two are
\itemize{
\item \code{aes(showSelected=variable)} means that
  only the subset of the data that corresponds to
  the selected value of variable will be shown.
\item \code{aes(clickSelects=variable)} means that clicking
  this geom will change the currently selected value of variable.
}
The others are described on https://github.com/tdhock/animint/wiki/Advanced-features-present-animint-but-not-in-ggplot2
}
\details{
Supported ggplot2 geoms:
\itemize{
\item point
\item jitter
\item line
\item rect
\item tallrect (new with this package)
\item segment
\item hline
\item vline
\item bar
\item text
\item tile
\item raster
\item ribbon
\item abline
\item density
\item path
\item polygon
\item histogram
\item violin
\item linerange
\item step
\item contour
\item density2d
\item area
\item freqpoly
\item hex
}
Unsupported geoms:
\itemize{
\item rug
\item dotplot
\item quantile - should *theoretically* work but in practice does not work
\item smooth - can be created using geom_line and geom_ribbon
\item boxplot - can be created using geom_rect and geom_segment
\item crossbar - can be created using geom_rect and geom_segment
\item pointrange - can be created using geom_linerange and geom_point
\item bin2d - bin using ddply() and then use geom_tile()
\item map - can be created using geom_polygon or geom_path
}
Supported scales:
\itemize{
\item alpha,
\item fill/colour (brewer, gradient, identity, manual)
\item linetype
\item x and y axis scales, manual break specification, label formatting
\item x and y axis theme elements: axis.line, axis.ticks, axis.text, axis.title can be set to element_blank(); other theme modifications not supported at this time, but would be possible with custom css files.
\item area
\item size
}
Unsupported scales:
\itemize{
\item shape. Open and closed circles can be represented by manipulating fill and colour scales and using default (circle) points, but d3 does not support many R shape types, so mapping between the two is difficult.
}
}
\seealso{
\code{\link{ggplot2}}
}
