% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimintproto.r
\name{as.list.gganimintproto}
\alias{as.list.gganimintproto}
\title{Convert a gganimintproto object to a list}
\usage{
\method{as.list}{gganimintproto}(x, inherit = TRUE, ...)
}
\arguments{
\item{x}{A gganimintproto object to convert to a list.}

\item{inherit}{If \code{TRUE} (the default), flatten all inherited items into
the returned list. If \code{FALSE}, do not include any inherited items.}

\item{...}{Further arguments to pass to \code{as.list.environment}.}
}
\description{
This will not include the object's \code{super} member.
}
