% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{checkAnimationTimeVar}
\alias{checkAnimationTimeVar}
\title{Check animation variable for errors}
\usage{
checkAnimationTimeVar(timeVarList)
}
\arguments{
\item{timeVarList}{\code{plot.list$time} in \code{animint2dir} to check
for errors}
}
\value{
\code{NULL} :Stops with an error for invalid input
}
\description{
Check animation variable for errors
}
