% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{checkExtraParams}
\alias{checkExtraParams}
\title{Check \code{extra_params} argument for duplicates, non-named list}
\usage{
checkExtraParams(extra_params, a_aes_mapping)
}
\arguments{
\item{extra_params}{named list containing the details of showSelected
and clickSelects values of the layer}

\item{a_aes_mapping}{aesthetics mapping of the layer}
}
\value{
Modified \code{extra_params} list
}
\description{
Check \code{extra_params} argument for duplicates, non-named list
}
