% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{checkForSSandCSasAesthetics}
\alias{checkForSSandCSasAesthetics}
\title{Check if showSelected and clickSelects have been used as a_aesthetics
as in old syntax. If yes, raise error}
\usage{
checkForSSandCSasAesthetics(a_aesthetics, plot_name)
}
\arguments{
\item{a_aesthetics}{list. aesthetics mapping of the layer}

\item{plot_name}{character vector of the plot the layer belongs to}
}
\value{
\code{NULL} Throws error if used as aesthetics
}
\description{
Check if showSelected and clickSelects have been used as a_aesthetics
as in old syntax. If yes, raise error
}
