% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{checkPlotForAnimintExtensions}
\alias{checkPlotForAnimintExtensions}
\title{Performs error checking on the plot for animint extensions}
\usage{
checkPlotForAnimintExtensions(p, plot_name)
}
\arguments{
\item{p}{plot from \code{plot.list} to check for errors}

\item{plot_name}{plot name error check. Should be alphanumeric and should
begin with an alphabet}
}
\value{
\code{NULL} :Stops with an error for invalid input
}
\description{
Performs error checking on the plot for animint extensions
}
