% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{checkPlotList}
\alias{checkPlotList}
\title{Check plot.list for errors}
\usage{
checkPlotList(plot.list)
}
\arguments{
\item{plot.list}{from \code{animint2dir} to check for errors}
}
\value{
Throws an error for invalid values
}
\description{
Check that plot.list is a list and every a_element is named
}
