% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{checkSelectorNames}
\alias{checkSelectorNames}
\title{Validate selector names for CSS compatibility}
\usage{
checkSelectorNames(selectors)
}
\arguments{
\item{selectors}{selectors to validate}
}
\value{
\code{NULL}. Throws error if invalid characters found.
}
\description{
Validate selector names for CSS compatibility
}
