% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{checkSingleShowSelectedValue}
\alias{checkSingleShowSelectedValue}
\title{Issue warnings for non interactive plots where there is only one
showSelected value}
\usage{
checkSingleShowSelectedValue(selectors)
}
\arguments{
\item{selectors}{selectors to check for warnings}
}
\value{
\code{NULL}
}
\description{
Issue warnings for non interactive plots where there is only one
showSelected value
}
