% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{colsNotToCopy}
\alias{colsNotToCopy}
\title{Filter out columns that do not need to be copied}
\usage{
colsNotToCopy(g, s.a_aes)
}
\arguments{
\item{g}{Geom with columns}

\item{s.a_aes}{Selector aesthetics}
}
\value{
Character vector of columns not to be copied
}
\description{
Filter out columns that do not need to be copied
}
