% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{common_value_for_group_subset}
\alias{common_value_for_group_subset}
\title{Common values for one (column, group) across chunk.vars subsets.}
\usage{
common_value_for_group_subset(value_lists)
}
\arguments{
\item{value_lists}{list of atomic vectors, one per chunk.vars level.}
}
\value{
list with common (list of length 1) and is.common (logical).
}
\description{
Common values for one (column, group) across chunk.vars subsets.
}
\keyword{internal}
