% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-munch.r
\name{dist_central_angle}
\alias{dist_central_angle}
\title{Compute central angle between two points.
Multiple by radius of sphere to get great circle distance}
\usage{
dist_central_angle(lon, lat)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}
}
\description{
Compute central angle between two points.
Multiple by radius of sphere to get great circle distance
}
