% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend-draw.r
\name{draw_key}
\alias{draw_key}
\alias{a_draw_key_point}
\alias{a_draw_key_abline}
\alias{a_draw_key_rect}
\alias{a_draw_key_polygon}
\alias{a_draw_key_blank}
\alias{a_draw_key_boxplot}
\alias{a_draw_key_crossbar}
\alias{a_draw_key_path}
\alias{a_draw_key_vpath}
\alias{a_draw_key_dotplot}
\alias{a_draw_key_pointrange}
\alias{a_draw_key_smooth}
\alias{a_draw_key_text}
\alias{a_draw_key_label}
\alias{a_draw_key_vline}
\title{Key drawing functions}
\usage{
a_draw_key_point(data, params, size)

a_draw_key_abline(data, params, size)

a_draw_key_rect(data, params, size)

a_draw_key_polygon(data, params, size)

a_draw_key_blank(data, params, size)

a_draw_key_boxplot(data, params, size)

a_draw_key_crossbar(data, params, size)

a_draw_key_path(data, params, size)

a_draw_key_vpath(data, params, size)

a_draw_key_dotplot(data, params, size)

a_draw_key_pointrange(data, params, size)

a_draw_key_smooth(data, params, size)

a_draw_key_text(data, params, size)

a_draw_key_label(data, params, size)

a_draw_key_vline(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\value{
A grid grob.
}
\description{
Each Geom has an associated function that draws the key when the geom needs
to be displayed in a legend. These are the options built into ggplot2.
}
\keyword{internal}
