% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economics}
\alias{economics}
\alias{economics_long}
\title{US economic time series.}
\format{
A data frame with 478 rows and 6 variables
\itemize{
  \item date.  Month of data collection
  \item psavert, personal savings rate,
  \item pce, personal consumption expenditures, in billions of dollars,
  \item unemploy, number of unemployed in thousands,
  \item uempmed, median duration of unemployment, in week,
  \item pop, total population, in thousands,
}

An object of class \code{data.frame} with 2870 rows and 4 columns.
}
\usage{
economics

economics_long
}
\description{
This dataset was produced from US economic time series data available from
\url{https://fred.stlouisfed.org/}. \code{economics} is in "wide"
format, \code{economics_long} is in "long" format.
}
\keyword{datasets}
