% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-munch.r
\name{find_line_formula}
\alias{find_line_formula}
\title{Given n points, find the slope, xintercept, and yintercept of
the lines connecting them.}
\usage{
find_line_formula(x, y)
}
\arguments{
\item{x}{A vector of x values}

\item{y}{A vector of y values}
}
\description{
This returns a data frame with length(x)-1 rows
}
\examples{
find_line_formula(c(4, 7), c(1, 5))
find_line_formula(c(4, 7, 9), c(1, 5, 3))
}
