% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_facets.R
\name{flag_axis}
\alias{flag_axis}
\title{Attach AXIS_X/AXIS_Y columns to the panel layout if 
facet_grids is used.
Currently every axis is rendered,
but this could be helpful if we decide not to that.}
\usage{
flag_axis(a_facet, layout)
}
\arguments{
\item{a_facet}{....}

\item{layout}{...}
}
\description{
Attach AXIS_X/AXIS_Y columns to the panel layout if 
facet_grids is used.
Currently every axis is rendered,
but this could be helpful if we decide not to that.
}
