% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.r
\name{g_train_layout}
\alias{g_train_layout}
\title{Learn the layout of panels within a plot.}
\usage{
g_train_layout(panel, facet, data, plot_data)
}
\arguments{
\item{panel}{the panel object to train}

\item{facet}{the facetting specification}

\item{data}{a list of data frames (one for each layer), and one for the plot}

\item{plot_data}{the default data frame}
}
\value{
an updated panel object
}
\description{
This is determined by the facet, which returns a data frame, than
when joined to the data to be plotted tells us which panel it should
appear in, where that panel appears in the grid, and what scales it
uses.
}
\details{
As well as the layout info, this function also adds empty lists in which
to house the x and y scales.
}
\keyword{internal}
