% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-dotplot.r
\name{geom_dotplot}
\alias{geom_dotplot}
\title{Dot plot (removed)}
\usage{
geom_dotplot(...)
}
\description{
`geom_dotplot()` has been removed from animint2 due to fundamental issues
with interactive rendering. Please use `geom_point()` instead.
}
\keyword{internal}
