% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms.R
\name{geom_tallrect}
\alias{geom_tallrect}
\title{ggplot2 geom with xmin and xmax aesthetics that covers the entire y range, useful for clickSelects background elements.}
\usage{
geom_tallrect(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{data}{data set}

\item{stat}{statistic mapping, defaults to identity}

\item{position}{position mapping, defaults to identity}

\item{...}{other arguments}
}
\value{
ggplot2 layer
}
\description{
ggplot2 geom with xmin and xmax aesthetics that covers the entire y range, useful for clickSelects background elements.
}
