% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{getCommonChunk}
\alias{getCommonChunk}
\title{Save the common columns for each tsv to one chunk}
\usage{
getCommonChunk(built, chunk.vars, aes.list)
}
\arguments{
\item{built}{data.frame of built data.}

\item{chunk.vars}{which variables to chunk on.}

\item{aes.list}{a character vector of aesthetics.}

\item{vars}{character vector of chunk variable names to split on.}
}
\value{
a list of common and varied data to save, or NULL if there is
no common data.
}
\description{
Save the common columns for each tsv to one chunk
}
