% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{getLayerName}
\alias{getLayerName}
\title{Gives a unique name to each layer in \code{saveLayer}}
\usage{
getLayerName(L, geom_num, p.name)
}
\arguments{
\item{L}{layer in saveLayer to be named}

\item{geom_num}{the number of the layer to be saved}

\item{p.name}{the name of the plot to which the layer belongs}
}
\value{
a unique name for the layer
}
\description{
Gives a unique name to each layer in \code{saveLayer}
}
