% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{getLayerParams}
\alias{getLayerParams}
\title{Get all parameters for a layer}
\usage{
getLayerParams(l)
}
\arguments{
\item{l}{A single layer of the plot}
}
\value{
All parameters in the layer
}
\description{
Get all parameters for a layer
}
