% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{getLegend}
\alias{getLegend}
\title{Function to get legend information for each scale}
\usage{
getLegend(mb)
}
\arguments{
\item{mb}{single entry from a_guides_merge() list of legend data}
}
\value{
list of legend information, NULL if a_guide=FALSE.
}
\description{
Function to get legend information for each scale
}
