% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animint.R
\name{getLegendList}
\alias{getLegendList}
\title{Function to get legend information from a_plot}
\usage{
getLegendList(plistextra)
}
\arguments{
\item{plistextra}{output from a_plot_build(p)}
}
\value{
list containing information for each legend
}
\description{
Function to get legend information from a_plot
}
