% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_facets.R
\name{getStrips}
\alias{getStrips}
\title{Determine titles to put on a_facet panels.
The implementation here is a modified version of ggplot2:::facet_strips.}
\usage{
getStrips(a_facet, panel, ...)
}
\arguments{
\item{a_facet}{....}

\item{panel}{....}

\item{...}{other arguments}
}
\description{
Determine titles to put on a_facet panels.
The implementation here is a modified version of ggplot2:::facet_strips.
}
