% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{getTextSize}
\alias{getTextSize}
\title{Function to process text size with different types of unit}
\usage{
getTextSize(element.name, theme)
}
\arguments{
\item{element.name}{The name of the theme element}

\item{theme}{combined theme from plot_theme()}
}
\value{
character of text size, with unit pt/px
}
\description{
Function to process text size with different types of unit
}
