% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{getUniqueAxisLabels}
\alias{getUniqueAxisLabels}
\title{Get unique axis labels for the plot}
\usage{
getUniqueAxisLabels(plot.meta)
}
\arguments{
\item{plot.meta}{contains axis labels}
}
\value{
modified \code{plot.meta} with unique axis labels
}
\description{
Get unique axis labels for the plot
}
