% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.r
\name{ggplotGrob}
\alias{ggplotGrob}
\title{Generate a ggplot2 (a_plot) plot grob.}
\usage{
ggplotGrob(x)
}
\arguments{
\item{x}{a_plot2 object}
}
\description{
Generate a ggplot2 (a_plot) plot grob.
}
\keyword{internal}
