% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{guides_geom}
\alias{guides_geom}
\title{guides_geom function}
\usage{
guides_geom(gdefs, layers, default_mapping)
}
\arguments{
\item{gdefs}{...}

\item{layers}{....}

\item{default_mapping}{...}
}
\description{
guides_geom function
}
