% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-.r
\name{guides_train}
\alias{guides_train}
\title{train each scale in scales and generate the definition of guide}
\usage{
guides_train(scales, theme, guides, labels)
}
\arguments{
\item{scales}{...}

\item{theme}{...}

\item{guides}{...}

\item{labels}{....}
}
\description{
train each scale in scales and generate the definition of guide
}
