% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-summary.r
\name{hmisc}
\alias{hmisc}
\alias{a_mean_cl_boot}
\alias{a_mean_cl_normal}
\alias{a_mean_sdl}
\alias{a_median_hilow}
\title{Wrap up a selection of summary functions from Hmisc to make it easy to use
with \code{\link{a_stat_summary}}.}
\usage{
a_mean_cl_boot(x, ...)

a_mean_cl_normal(x, ...)

a_mean_sdl(x, ...)

a_median_hilow(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{other arguments passed on to the respective Hmisc function.}
}
\description{
See the Hmisc documentation for details of their options.
}
\seealso{
\code{\link[Hmisc]{smean.cl.boot}},
  \code{\link[Hmisc]{smean.cl.normal}}, \code{\link[Hmisc]{smean.sdl}},
   \code{\link[Hmisc]{smedian.hilow}}
}
