% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-.r
\name{is.a_facet}
\alias{is.a_facet}
\title{Is this object a facetting specification?}
\usage{
is.a_facet(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is this object a facetting specification?
}
\keyword{internal}
