% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{is.a_plot}
\alias{is.a_plot}
\title{Reports whether x is a a_plot object}
\usage{
is.a_plot(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
Reports whether x is a a_plot object
}
\keyword{internal}
