% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimintproto.r
\name{is.gganimintproto}
\alias{is.gganimintproto}
\title{Is an object a gganimintproto object?}
\usage{
is.gganimintproto(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Is an object a gganimintproto object?
}
