% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{issueSelectorWarnings}
\alias{issueSelectorWarnings}
\title{Issue warnings for selectors}
\usage{
issueSelectorWarnings(geoms, selector.a_aes, duration)
}
\arguments{
\item{geoms}{\code{geoms} to check for warnings}

\item{selector.a_aes}{selectors for each geom}

\item{duration}{animation variable information to check for \code{key} value}
}
\value{
\code{NULL}
}
\description{
Issue warnings for selectors
}
