% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-layout.r
\name{layout_base}
\alias{layout_base}
\title{Base layout function that generates all combinations of data needed for
facetting
The first data frame in the list should be the default data for the plot.
Other data frames in the list are ones that are added to layers.}
\usage{
layout_base(data, vars = NULL, drop = TRUE)
}
\arguments{
\item{data}{list of data frames (one for each layer)}
}
\description{
Base layout function that generates all combinations of data needed for
facetting
The first data frame in the list should be the default data for the plot.
Other data frames in the list are ones that are added to layers.
}
\keyword{internal}
