% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-layout.r
\name{layout_grid}
\alias{layout_grid}
\title{Layout panels in a 2d grid.}
\usage{
layout_grid(data, rows = NULL, cols = NULL, margins = NULL, drop = TRUE,
  as.table = TRUE)
}
\arguments{
\item{data}{list of data frames, one for each layer}

\item{rows}{variables that form the rows}

\item{cols}{variables that form the columns}

\item{margins}{......}

\item{drop}{....}

\item{as.table}{....}
}
\value{
a data frame with columns \code{PANEL}, \code{ROW} and \code{COL},
  that match the facetting variable values up with their position in the
  grid
}
\description{
Layout panels in a 2d grid.
}
