% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet-layout.r
\name{layout_wrap}
\alias{layout_wrap}
\title{Layout out panels in a 1d ribbon.}
\usage{
layout_wrap(data, vars = NULL, nrow = NULL, ncol = NULL,
  as.table = TRUE, drop = TRUE, dir = "h")
}
\arguments{
\item{drop}{should missing combinations be excluded from the plot?}
}
\description{
Layout out panels in a 1d ribbon.
}
\keyword{internal}
