% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.r
\name{map_layout}
\alias{map_layout}
\title{Map data to find out where it belongs in the plot.}
\usage{
map_layout(panel, a_facet, data)
}
\arguments{
\item{panel}{a trained panel object}

\item{data}{list of data frames (one for each layer)}

\item{the}{facetting specification}
}
\description{
Layout map ensures that all layer data has extra copies of data for margins
and missing facetting variables, and has a PANEL variable that tells that
so it know what panel it belongs to. This is a change from the previous
design which added facetting variables directly to the data frame and
caused problems when they had names of aesthetics (like colour or group).
}
\keyword{internal}
