% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.r
\name{map_position}
\alias{map_position}
\title{Map data with scales.}
\usage{
map_position(panel, data, x_scale, y_scale)
}
\arguments{
\item{data}{a list of data frames (one for each layer)}
}
\description{
This operation must be idempotent because it is applied twice: both before
and after statistical transformation.
}
\keyword{internal}
