\name{montreal.bikes}
\alias{montreal.bikes}
\docType{data}
\title{
  Montreal bikes
}
\description{
  Data on montreal bike usage, paths, and accidents, 2009-2014.
}
\usage{data("montreal.bikes")}
\format{A named list of four data.frames. counter.counts contains dates
  and counts of bikes at several locations in
  Montreal. counter.locations has the geographical coordinates of the
  counter locations. path.locations has the geographical coordinates of
  several Montreal bike paths. accidents has
  dates/times/locations/severity of accidents involving bikes.
}
\source{
  github.com/tdhock/montreal-velos
}
\keyword{datasets}
