% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-munch.r
\name{munch_data}
\alias{munch_data}
\title{For munching, only grobs are lines and polygons: everything else is
transformed into those special cases by the geom.}
\usage{
munch_data(data, dist = NULL, segment_length = 0.01)
}
\arguments{
\item{dist}{distance, scaled from 0 to 1 (maximum distance on plot)}
}
\description{
For munching, only grobs are lines and polygons: everything else is
transformed into those special cases by the geom.
}
\keyword{internal}
