% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{newEnvironment}
\alias{newEnvironment}
\title{Environment to store meta data}
\usage{
newEnvironment()
}
\value{
A new environment to store meta data
}
\description{
Get a new environment to store meta-data. Used to alter state in the
lower-level functions
}
