% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animint.R
\name{parsePlot}
\alias{parsePlot}
\title{Convert a ggplot to a list.}
\usage{
parsePlot(meta, plot, plot.name)
}
\arguments{
\item{meta}{environment with previously calculated plot data, and a new plot to parse, already stored in plot and plot.name.}

\item{plot}{ggplot list object}

\item{plot.name}{name of plot}
}
\value{
nothing, info is stored in meta.
}
\description{
Convert a ggplot to a list.
}
