% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{plot_a_theme}
\alias{plot_a_theme}
\title{Combine plot defaults with current a_theme to get complete a_theme for a plot}
\usage{
plot_a_theme(x)
}
\arguments{
\item{x}{....}
}
\description{
Combine plot defaults with current a_theme to get complete a_theme for a plot
}
