% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_print.R
\name{print.animint}
\alias{print.animint}
\title{print animint}
\usage{
\method{print}{animint}(x, ...)
}
\arguments{
\item{x}{List of ggplots and options}

\item{...}{passed to animint2dir}
}
\value{
same as animint2dir
}
\description{
Print animint by rendering to local directory.
}
\author{
Toby Dylan Hocking
}
