% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{print.gganimintplot}
\alias{print.gganimintplot}
\alias{plot.gganimintplot}
\title{Draw plot on current graphics device.}
\usage{
\method{print}{gganimintplot}(x, newpage = is.null(vp), vp = NULL, ...)

\method{plot}{gganimintplot}(x, newpage = is.null(vp), vp = NULL, ...)
}
\arguments{
\item{x}{plot to display}

\item{newpage}{draw new (empty) page first?}

\item{vp}{viewport to draw plot in}

\item{...}{other arguments not used by this method}
}
\value{
Invisibly returns the result of \code{\link{ggplot_build}}, which
  is a list with components that contain the plot itself, the data,
  information about the scales, panels etc.
}
\description{
Draw plot on current graphics device.
}
\keyword{hplot}
