% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gganimintproto.r
\name{print.gganimintproto}
\alias{print.gganimintproto}
\title{Print a gganimintproto object}
\usage{
\method{print}{gganimintproto}(x, ..., flat = TRUE)
}
\arguments{
\item{x}{A gganimintproto object to print.}

\item{...}{If the gganimintproto object has a \code{print} method, further arguments
will be passed to it. Otherwise, these arguments are unused.}

\item{flat}{If \code{TRUE} (the default), show a flattened list of all local
and inherited members. If \code{FALSE}, show the inheritance hierarchy.}
}
\description{
If a gganimintproto object has a \code{$print} method, this will call that method.
Otherwise, it will print out the members of the object, and optionally, the
members of the inherited objects.
}
