% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_helperFunctions.R
\name{pt.to.lines}
\alias{pt.to.lines}
\title{Convert \code{pt} value to \code{lines}}
\usage{
pt.to.lines(pt_value)
}
\arguments{
\item{pt_value}{Value in \code{pt} to be converted to \code{lines}}
}
\value{
Value in \code{lines}
}
\description{
Convert \code{pt} value to \code{lines}
}
\note{
Does NOT work if input is not in \code{pt}. Input is returned as is.
}
