% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r, R/utilities-grid.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{alpha}
\alias{reexports}
\alias{unit}
\alias{reexports}
\alias{arrow}
\title{Objects exported from other packages}
\examples{
a_plot(mpg, a_aes(displ, hwy)) +
  a_geom_point(alpha = 0.5, colour = "blue")

a_plot(mpg, a_aes(displ, hwy)) +
  a_geom_point(colour = alpha("blue", 0.5))
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grid}{\code{\link[grid]{unit}}, \code{\link[grid]{arrow}}}

  \item{scales}{\code{\link[scales]{alpha}}}
}}

