% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animint.R
\name{saveLayer}
\alias{saveLayer}
\title{Save a layer to disk, save and return meta-data.}
\usage{
saveLayer(l, d, meta, layer_name, ggplot, built, AnimationInfo)
}
\arguments{
\item{l}{one layer of the ggplot object.}

\item{d}{one layer of calculated data from ggplot_build(p).}

\item{meta}{environment of meta-data.}

\item{layer_name}{name of layer}

\item{ggplot}{ggplot}

\item{built}{built list}

\item{AnimationInfo}{animation list
ID number starting from 1}
}
\value{
list representing a layer, with corresponding aesthetics, ranges, and groups.
}
\description{
Save a layer to disk, save and return meta-data.
}
