% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{selectSSandCS}
\alias{selectSSandCS}
\title{Separate .variable/.value selectors}
\usage{
selectSSandCS(aesthetics_list)
}
\arguments{
\item{aesthetics_list}{aesthetics mapping of the layer}
}
\value{
Modified \code{aes.list} list with separated
showSelected.variable/value
}
\description{
Separate .variable/.value selectors
}
