% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{setPlotSizes}
\alias{setPlotSizes}
\title{Set plot width and height for all plots}
\usage{
setPlotSizes(meta, AllPlotsInfo)
}
\arguments{
\item{meta}{meta object with all information}

\item{AllPlotsInfo}{plot info list}
}
\value{
\code{NULL}. Sizes are stored in meta object
}
\description{
Set plot width and height for all plots
}
