% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{split_recursive}
\alias{split_recursive}
\title{Split data.frame into recursive list of data.frame.}
\usage{
split_recursive(x, vars)
}
\arguments{
\item{x}{data.frame.}

\item{vars}{character vector of variable names to split on.}
}
\value{
recursive list of data.frame.
}
\description{
Split data.frame into recursive list of data.frame.
}
