% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-path.r
\name{stairstep}
\alias{stairstep}
\title{Calculate stairsteps
Used by \code{\link{a_geom_step}}}
\usage{
stairstep(data, direction = "hv")
}
\arguments{
\item{data}{...}

\item{direction}{...}
}
\description{
Calculate stairsteps
Used by \code{\link{a_geom_step}}
}
