% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.r
\name{summary.a_plot}
\alias{summary.a_plot}
\title{Displays a useful description of a ggplot object}
\usage{
\method{summary}{a_plot}(object, ...)
}
\arguments{
\item{object}{ggplot2 object to summarise}

\item{...}{other arguments ignored (for compatibility with generic)}
}
\description{
Displays a useful description of a ggplot object
}
\examples{
p <- a_plot(mtcars, a_aes(mpg, wt)) +
  a_geom_point()
summary(p)
}
\keyword{internal}
