% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{switch_axes}
\alias{switch_axes}
\title{Flip axes in case of a_coord_flip
Switches column names. Eg. xmin to ymin, yntercept to xintercept etc.}
\usage{
switch_axes(col.names)
}
\arguments{
\item{col.names}{Column names which need to be switched}
}
\value{
Column names with x and y axes switched
}
\description{
Flip axes in case of a_coord_flip
Switches column names. Eg. xmin to ymin, yntercept to xintercept etc.
}
