\name{TestROC}
\alias{TestROC}
\docType{data}
\title{
Test ROC curves
}
\description{
Five peak detection models were evaluated.
}
\usage{data("TestROC")}
\format{
  TODO add codebook
  
  A list of two data frames.
}
\source{
  https://bitbucket.org/mugqic/chip-seq-paper/src/master/figure-roc-test.R
}
\keyword{datasets}
