% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.r
\name{train_position}
\alias{train_position}
\title{Train position scales with data}
\usage{
train_position(panel, data, x_scale, y_scale)
}
\arguments{
\item{panel}{the panel object to train}

\item{data}{a list of data frames (one for each layer)}

\item{x_scale}{x scale for the plot}

\item{y_scale}{y scale for the plot}
}
\description{
If panel-specific scales are not already present, will clone from
the scales provided in the parameter
}
\keyword{internal}
