% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.r
\name{update_a_labels}
\alias{update_a_labels}
\title{Update axis/legend labels}
\usage{
update_a_labels(p, a_labels)
}
\arguments{
\item{p}{plot to modify}

\item{a_labels}{named list of new a_labels}
}
\description{
Update axis/legend labels
}
\examples{
p <- a_plot(mtcars, a_aes(mpg, wt)) + a_geom_point()
update_a_labels(p, list(x = "New x"))
update_a_labels(p, list(x = expression(x / y ^ 2)))
update_a_labels(p, list(x = "New x", y = "New Y"))
update_a_labels(p, list(colour = "Fail silently"))
}
