% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{update_a_theme}
\alias{update_a_theme}
\title{Update a a_theme from a plot object}
\usage{
update_a_theme(olda_theme, newa_theme)
}
\arguments{
\item{olda_theme}{an existing a_theme, usually from a plot object, like
plot$a_theme. This could be an empty list.}

\item{newa_theme}{a new a_theme object to add to the existing a_theme}
}
\description{
This is called from add_a_plot.
}
\details{
If newa_theme is a *complete* a_theme, then it is meant to replace
olda_theme; this function just returns newa_theme.

Otherwise, it adds elements from newa_theme to olda_theme:
If olda_theme doesn't already contain those elements,
it searches the current default a_theme, grabs the elements with the
same name as those from newa_theme, and puts them in olda_theme. Then
it adds elements from newa_theme to olda_theme.
This makes it possible to do things like:
  a_plot(data.frame(x = 1:3, y = 1:3)) +
  a_geom_point() + a_theme(text = a_element_text(colour = 'red'))
and have 'text' keep properties from the default a_theme. Otherwise
you would have to set all the element properties, like family, size,
etc.
}
