% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_pages.R
\name{update_gallery}
\alias{update_gallery}
\title{Update gallery}
\usage{
update_gallery(gallery_path = "~/R/gallery")
}
\arguments{
\item{gallery_path}{path to local github repo with gh-pages
active.}
}
\value{
named list of data tables (meta and error).
}
\description{
A gallery is a collection of animints that have been published to
github pages. First repos.txt is read, then we clone each repo
which is not already present in meta.csv, and parse meta-data
(title, source, Capture.PNG) from the gh-pages branch, and
write/commit the data, re-render index.Rmd in gallery, and push
gallery to origin. See
\url{https://github.com/animint/gallery/tree/gh-pages} which is
the main gallery which is updated using this function.
}
\author{
Toby Dylan Hocking
}
