% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.r
\name{update_theme}
\alias{update_theme}
\title{Update a theme from a plot object}
\usage{
update_theme(oldtheme, newtheme)
}
\arguments{
\item{oldtheme}{an existing theme, usually from a plot object, like
plot$theme. This could be an empty list.}

\item{newtheme}{a new theme object to add to the existing theme}
}
\description{
This is called from add_a_plot.
}
\details{
If newtheme is a *complete* theme, then it is meant to replace
oldtheme; this function just returns newtheme.

Otherwise, it adds elements from newtheme to oldtheme:
If oldtheme doesn't already contain those elements,
it searches the current default theme, grabs the elements with the
same name as those from newtheme, and puts them in oldtheme. Then
it adds elements from newtheme to oldtheme.
This makes it possible to do things like:
  a_plot(data.frame(x = 1:3, y = 1:3)) +
  geom_point() + theme(text = a_element_text(colour = 'red'))
and have 'text' keep properties from the default theme. Otherwise
you would have to set all the element properties, like family, size,
etc.
}
