% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animintHelpers.R
\name{varied.chunk}
\alias{varied.chunk}
\title{Extract subset for each data.frame in a list of data.frame}
\usage{
varied.chunk(dt.or.list, cols)
}
\arguments{
\item{dt.or.list}{a data.table or a list of data.table.}

\item{cols}{cols that each data.frame would keep.}
}
\value{
list of data.frame.
}
\description{
Extract subset for each data.frame in a list of data.frame
}
