% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_print.R
\name{animint}
\alias{animint}
\title{create an animint}
\usage{
animint(...)
}
\arguments{
\item{...}{ggplots and options}
}
\value{
list of class animint
}
\description{
create a list and attach animint class
}
\author{
Toby Dylan Hocking
}
