% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animint.R
\name{animint2dir}
\alias{animint2dir}
\alias{animint}
\title{Compile and render an animint in a local directory}
\usage{
animint2dir(plot.list, out.dir = tempfile(), json.file = "plot.json",
  open.browser = interactive(), css.file = "")
}
\arguments{
\item{plot.list}{a named list of ggplots and option lists.}

\item{out.dir}{directory to store html/js/csv files.}

\item{json.file}{character string that names the JSON file with metadata associated with the plot.}

\item{open.browser}{Should R open a browser? If yes, be sure to configure your browser to allow access to local files, as some browsers block this by default (e.g. chrome).}

\item{css.file}{character string for non-empty css file to include. Provided file will be copied to the output directory as styles.css}
}
\value{
invisible list of ggplots in list format.
}
\description{
An animint is a list of ggplots and options that defines
an interactive animation and can be viewed in a web browser.
Several new aesthetics control interactivity.
The most important two are
\itemize{
\item \code{a_aes(showSelected=variable)} means that
  only the subset of the data that corresponds to
  the selected value of variable will be shown.
\item \code{a_aes(clickSelects=variable)} means that clicking
  this geom will change the currently selected value of variable.
}
The others are described on https://github.com/tdhock/animint/wiki/Advanced-features-present-animint-but-not-in-ggplot2
}
\details{
Supported ggplot2 geoms:
\itemize{
\item point
\item jitter
\item line
\item rect
\item tallrect (new with this package)
\item segment
\item hline
\item vline
\item bar
\item text
\item tile
\item raster
\item ribbon
\item abline
\item density
\item path
\item polygon
\item histogram
\item violin
\item linerange
\item step
\item contour
\item density2d
\item area
\item freqpoly
\item hex
}
Unsupported geoms:
\itemize{
\item rug
\item dotplot
\item quantile - should *theoretically* work but in practice does not work
\item smooth - can be created using a_geom_line and a_geom_ribbon
\item boxplot - can be created using a_geom_rect and a_geom_segment
\item crossbar - can be created using a_geom_rect and a_geom_segment
\item pointrange - can be created using a_geom_linerange and a_geom_point
\item bin2d - bin using ddply() and then use a_geom_tile()
\item map - can be created using a_geom_polygon or a_geom_path
}
Supported scales:
\itemize{
\item alpha,
\item fill/colour (brewer, gradient, identity, manual)
\item linetype
\item x and y axis scales, manual break specification, a_label formatting
\item x and y axis a_theme elements: axis.line, axis.ticks, axis.text, axis.title can be set to a_element_blank(); other a_theme modifications not supported at this time, but would be possible with custom css files.
\item area
\item size
}
Unsupported scales:
\itemize{
\item shape. Open and closed circles can be represented by manipulating fill and colour scales and using default (circle) points, but d3 does not support many R shape types, so mapping between the two is difficult.
}
}
\examples{
## Make a Gapminder plot (aka Google motion chart), which is actually
## just a scatterplot with size and color that moves over time.
library(animint2)
data(WorldBank)
gapminder <-
  list(title="Linked scatterplot and time series",
       ts=a_plot()+
       make_tallrect(WorldBank, "year")+
       a_geom_line(a_aes(year, life.expectancy, group=country, color=region),
                 clickSelects="country",
                 data=WorldBank, size=4, alpha=3/5),
       time=list(variable="year",ms=3000),
       duration=list(year=1000),
       scatter=a_plot()+
       a_geom_point(a_aes(fertility.rate, life.expectancy,
                      key=country, colour=region, size=population),
                  showSelected="year",
                  clickSelects="country",
                  data=WorldBank)+
       a_geom_text(a_aes(fertility.rate, life.expectancy, a_label=country),
                 showSelected=c("country", "year"),
                 data=WorldBank)+
       make_text(WorldBank, 5, 80, "year")+
       a_scale_size_animint(pixel.range=c(2,20), breaks=10^(4:9)))
animint2dir(gapminder, "WorldBank-viz")

data(worldPop)
## Linked bar and line plots of world population by subcontinent,
## inspired by polychartjs.
popPlots <-
  list(bars=a_plot()+
       a_geom_bar(a_aes(x=subcontinent, y=population),
                clickSelects="subcontinent",
                showSelected="year",
                data=worldPop, a_stat="identity", a_position="identity")+
       ## This make_text creates a a_geom_text that shows the current
       ## selected value of the year variable.
       make_text(worldPop, 1, 3e6, "year")+
       a_coord_flip(),
       lines=a_plot()+
       ## This make_tallrect tiles the background of the lineplot with
       ## rects that can be clicked to select the year variable.
       make_tallrect(worldPop, "year")+
       ## This a_geom_point does not have a_aes(clickSelects) so its alpha
       ## transparency behaves normally: all points have alpha=1/4.
       a_geom_point(a_aes(year, population, colour=type),
                  data=worldPop, size=4, alpha=1/4)+
       ## This a_geom_line DOES have a_aes(clickSelects) so only the
       ## selected line has the specified alpha=3/4. The other
       ## unselected lines have 0.5 less (alpha=1/4).
       a_geom_line(a_aes(year, population, group=subcontinent),
                 clickSelects="subcontinent",
                 data=worldPop, size=4, alpha=3/4))
animint2dir(popPlots, "WorldPop-interactive")
## Make it animated by specifying year as the variable to animate and
## an interval of 2000 milliseconds between animation frames.
popAnim <- c(popPlots, list(time=list(variable="year",ms=2000)))
animint2dir(popAnim, "WorldPop-animated")
## Make the animation smooth by specifying a duration of 1000 ms for
## geoms with a_aes(showSelected=year).
popSmooth <- c(popAnim, list(duration=list(year=1000)))
animint2dir(popSmooth, "WorldPop-smooth")

}
