% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guides-axis.r
\name{guide_axis}
\alias{guide_axis}
\title{Grob for axes}
\usage{
guide_axis(at, labels, a_position = "right", a_theme)
}
\arguments{
\item{at}{...}

\item{labels}{at ticks}

\item{a_position}{of ticks}

\item{a_theme}{...}
}
\description{
Grob for axes
}
