% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_animint.R
\name{saveLayer}
\alias{saveLayer}
\title{Save a layer to disk, save and return meta-data.}
\usage{
saveLayer(l, d, meta, a_layer_name, a_plot, built, AnimationInfo)
}
\arguments{
\item{l}{one layer of the ggplot object.}

\item{d}{one layer of calculated data from a_plot_build(p).}

\item{meta}{environment of meta-data.}

\item{a_layer_name}{.....}

\item{a_plot}{....}

\item{built}{....}

\item{geom_num}{the number of geom in the plot. Each geom gets an increasing}

\item{AniminationInfo}{ID number starting from 1}
}
\value{
list representing a layer, with corresponding aesthetics, ranges, and groups.
}
\description{
Save a layer to disk, save and return meta-data.
}
