% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-defaults.r
\name{update_a_geom_defaults}
\alias{update_a_geom_defaults}
\alias{update_a_stat_defaults}
\title{Modify geom/stat aesthetic defaults for future plots}
\usage{
update_a_geom_defaults(a_geom, new)

update_a_stat_defaults(a_stat, new)
}
\arguments{
\item{new}{Named list of aesthetics.}

\item{a_stat, a_geom}{Name of a_geom/a_stat to modify (like \code{"point"} or
\code{"bin"}), or a a_Geom/a_Stat object (like \code{a_GeomPoint} or
\code{a_StatBin}).}
}
\description{
Modify geom/stat aesthetic defaults for future plots
}
\examples{
update_a_geom_defaults("point", list(colour = "darkblue"))
a_plot(mtcars, a_aes(mpg, wt)) + a_geom_point()
update_a_geom_defaults("point", list(colour = "black"))
}
